% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{batch.info}
\alias{batch.info}
\title{TCGA batch information from Biospecimen Metadata Browser}
\format{A data frame with 11382 rows and 3 variables}
\description{
TCGA batch information from Biospecimen Metadata Browser
}
\keyword{internal}
