% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{bcgsc.ca_CHOL.IlluminaHiSeq_DNASeq.1.somatic.maf}
\alias{bcgsc.ca_CHOL.IlluminaHiSeq_DNASeq.1.somatic.maf}
\title{TCGA CHOL MAF}
\format{A tibble: 3,555 x 34}
\description{
TCGA CHOL MAF
}
\keyword{internal}
