% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{BRCA_rnaseqv2}
\alias{BRCA_rnaseqv2}
\title{BRCA_rnaseqv2}
\format{A data frame with 200 rows (genes) and 1172 variables (samples)}
\description{
BRCA_rnaseqv2
}
\keyword{internal}

