% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{BRCA_RNASeqV2_version}
\alias{BRCA_RNASeqV2_version}
\title{TCGA Table with version, number of samples and size (Mbyte) of
BRCA IlluminaHiSeq_RNASeqV2 Level 3}
\format{A data frame with 12 rows and 4 variables}
\description{
TCGA Table with version, number of samples and size (Mbyte) of
BRCA IlluminaHiSeq_RNASeqV2 Level 3
}
\keyword{internal}

