% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{calculate.pvalues}
\alias{calculate.pvalues}
\title{Calculate pvalues}
\usage{
calculate.pvalues(values, idx1, idx2, paired = TRUE, exact = TRUE,
  cores = NULL)
}
\arguments{
\item{values}{Dataframe with values}

\item{idx1}{Index of the values in group 1}

\item{idx2}{Index of the values in group 2}

\item{paired}{Do a paired wilcoxon test? Default: True}

\item{exact}{Do a exact wilcoxon test? Default: True}

\item{cores}{How many cores to be used Default: parallel::detectCores()}
}
\value{
Data frame with cols p values/p values adjusted

Data frame with two cols
        p-values/p-values adjusted
}
\description{
Calculate pvalues
}
\details{
Input: two matrix to be compared with wilcoxon test,
          a boolean value to do a paired or non-paired test
   Output: p-values (non-adj/adj) histograms, p-values (non-adj/adj)
}

