% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{classification}
\alias{classification}
\title{Result of gliomaclassifier function}
\format{
A list of data frames
}
\description{
Result of gliomaclassifier function
}
\keyword{internal}
