% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{clinic}
\alias{clinic}
\title{clinic}
\usage{
clinic(cancer, clinical_data_type)
}
\arguments{
\item{cancer}{cancer}

\item{clinical_data_type}{clinical_data_type}
}
\value{
clinic
}
\description{
clinic
}
\examples{
clinic ("gbm","clinical_patient")
}

