% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{clinical.biotab}
\alias{clinical.biotab}
\title{A list of data frames with clinical data parsed from XML (code in vignettes)}
\format{A list with 7 elements}
\description{
A list of data frames with clinical data parsed from XML (code in vignettes)
}
\keyword{internal}
