% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{clinical_data_site_cancer}
\alias{clinical_data_site_cancer}
\title{clinical_data_site_cancer}
\usage{
clinical_data_site_cancer(cancer)
}
\arguments{
\item{cancer}{cancer}
}
\value{
clinical_data_site_cancer
}
\description{
clinical_data_site_cancer
}
\keyword{internal}

