% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{clinical_data_site_cancer_test}
\alias{clinical_data_site_cancer_test}
\title{clinical_data_site_cancer_test}
\usage{
clinical_data_site_cancer_test(cancer)
}
\arguments{
\item{cancer}{cancer}
}
\value{
clinical_data_site_cancer_test
}
\description{
clinical_data_site_cancer_test
}

