% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{clinicFilt}
\alias{clinicFilt}
\title{Filter samples using clinical data}
\usage{
clinicFilt(barcode, clinical_patient_data, HER = NULL, ER = NULL,
  gender = NULL, PR = NULL, stage = NULL)
}
\arguments{
\item{barcode}{List of barcodes}

\item{clinical_patient_data}{clinical_patient_data obtained with clinic function
Ex: clinical_patient_data <- TCGAquery_clinic("LGG","clinical_patient")}

\item{HER}{her2 neu immunohistochemistry receptor status: "Positive" or "Negative"}

\item{ER}{Estrogen receptor status: "Positive" or "Negative"}

\item{gender}{"MALE" or "FEMALE"}

\item{PR}{Progesterone receptor status: "Positive" or "Negative"}

\item{stage}{Pathologic Stage: "stage_IX", "stage_I", "stage_IA", "stage_IB", "stage_IIX",
"stage_IIA", "stage_IIB", "stage_IIIX","stage_IIIA", "stage_IIIB",
"stage_IIIC", "stage_IV" -}
}
\value{
List of samples that matches the filters
}
\description{
This function will return the samples that matches all filters.
  Filters available: HER, ER,gender,PR, stage.
}
\examples{
# clin <- TCGAquery_clinic("BRCA","clinical_patient")
clin <- clinBRCA
bar <- c("TCGA-G9-6378-02A-11R-1789-07", "TCGA-CH-5767-04A-11R-1789-07",
        "TCGA-G9-6332-60A-11R-1789-07", "TCGA-G9-6336-01A-11R-1789-07",
        "TCGA-G9-6336-11A-11R-1789-07", "TCGA-G9-7336-11A-11R-1789-07",
        "TCGA-G9-7336-04A-11R-1789-07", "TCGA-G9-7336-14A-11R-1789-07",
        "TCGA-G9-7036-04A-11R-1789-07", "TCGA-G9-7036-02A-11R-1789-07",
        "TCGA-G9-7036-11A-11R-1789-07", "TCGA-G9-7036-03A-11R-1789-07",
        "TCGA-G9-7036-10A-11R-1789-07", "TCGA-BH-A1ES-10A-11R-1789-07",
        "TCGA-BH-A1F0-10A-11R-1789-07", "TCGA-BH-A0BZ-02A-11R-1789-07",
        "TCGA-B6-A0WY-04A-11R-1789-07", "TCGA-BH-A1FG-04A-11R-1789-08",
        "TCGA-D8-A1JS-04A-11R-2089-08", "TCGA-AN-A0FN-11A-11R-8789-08",
        "TCGA-AR-A2LQ-12A-11R-8799-08", "TCGA-AR-A2LH-03A-11R-1789-07",
        "TCGA-BH-A1F8-04A-11R-5789-07", "TCGA-AR-A24T-04A-55R-1789-07",
        "TCGA-AO-A0J5-05A-11R-1789-07", "TCGA-BH-A0B4-11A-12R-1789-07",
        "TCGA-B6-A1KN-60A-13R-1789-07", "TCGA-AO-A0J5-01A-11R-1789-07",
        "TCGA-AO-A0J5-01A-11R-1789-07", "TCGA-G9-6336-11A-11R-1789-07",
        "TCGA-G9-6380-11A-11R-1789-07", "TCGA-G9-6380-01A-11R-1789-07",
        "TCGA-G9-6340-01A-11R-1789-07","TCGA-G9-6340-11A-11R-1789-07")

clinicFilt(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),clin,
HER="Positive", gender="FEMALE",ER="Positive")
}

