% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{colDataPrepare}
\alias{colDataPrepare}
\title{Create samples information matrix for GDC samples}
\usage{
colDataPrepare(barcode)
}
\arguments{
\item{barcode}{TCGA or TARGET barcode}
}
\description{
Create samples information matrix for GDC samples add subtype information
}
\examples{
 metadata <- colDataPrepare(c("TCGA-OR-A5K3-01A","C3N-00321-01"))
 metadata <- colDataPrepare(c("BLGSP-71-06-00157-01A",
                              "BLGSP-71-22-00332-01A"))
}
