% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGA_biolinks_functions.R
\name{createDirectory}
\alias{createDirectory}
\title{Creates a directory}
\usage{
createDirectory(base)
}
\arguments{
\item{base}{directory path}
}
\description{
Internal use only.
}
\keyword{internal}

