% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{CreateTabLevel}
\alias{CreateTabLevel}
\title{CreateTabLevel for Differentially expression analysis (DEA)}
\usage{
CreateTabLevel(FC_FDR_table_mRNA, typeCond1, typeCond2, TableCond1, TableCond2,
  typeOrder = T)
}
\arguments{
\item{FC_FDR_table_mRNA}{FC_FDR_table_mRNA}

\item{typeCond1}{typeCond1}

\item{typeCond2}{typeCond2}

\item{TableCond1}{TableCond1}

\item{TableCond2}{TableCond2}

\item{typeOrder}{typeOrder}
}
\value{
table with DEGs (diff.expr. genes)
}
\description{
CreateTabLevel for Differentially expression analysis (DEA)
}

