% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{dataBRCA}
\alias{dataBRCA}
\title{TCGA data matrix BRCA}
\format{A data frame with 20531 rows (genes) and 50 variables (samples)}
\description{
TCGA data matrix BRCA
}
\keyword{internal}

