% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{dataDEGsFiltLevel}
\alias{dataDEGsFiltLevel}
\title{TCGA data matrix BRCA DEGs}
\format{A data frame with 3649 rows and 6 variables}
\description{
TCGA data matrix BRCA DEGs
}
\keyword{internal}

