% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{dataREAD}
\alias{dataREAD}
\title{TCGA data SummarizedExperiment READ}
\format{A SummarizedExperiment of READ with 2 samples}
\description{
TCGA data SummarizedExperiment READ
}
\keyword{internal}

