% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{dataREAD_df}
\alias{dataREAD_df}
\title{TCGA data matrix READ}
\format{A data frame with 20531 rows (genes) and 2 variables (samples)}
\description{
TCGA data matrix READ
}
\keyword{internal}

