% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{DAVID_MF_matrix}
\alias{DAVID_MF_matrix}
\title{DAVID GO table (MF)}
\format{A data frame with 1157 rows and 2 variables}
\description{
DAVID GO table (MF)
}
\keyword{internal}

