% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{DEArnaSEQ}
\alias{DEArnaSEQ}
\title{Differentially expression analysis (DEA)}
\usage{
DEArnaSEQ(mat1, mat2, Cond1type, Cond2type)
}
\arguments{
\item{mat1}{datamatrix normal}

\item{mat2}{datamatrix tumor}

\item{Cond1type}{normal}

\item{Cond2type}{tumor}
}
\value{
table with DEGs (diff.expr. genes)
}
\description{
Differentially expression analysis (DEA) with gene expression
}

