% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{diffmean}
\alias{diffmean}
\title{calculate diffmean methylation between two groups}
\usage{
diffmean(group1, group2)
}
\arguments{
\item{group1}{Data frame probes vs patient}

\item{group2}{Data frame probes vs patient}
}
\value{
dataframe with diffmean values
}
\description{
calculate diffmean methylation between two groups
}

