% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{EAbarplot}
\alias{EAbarplot}
\title{EAbarplot - EAcomplete plot}
\usage{
EAbarplot(tf, GOMFTab, GOBPTab, GOCCTab, PathTab, nBar, nRGTab)
}
\arguments{
\item{tf}{tf}

\item{GOMFTab}{GOMFTab}

\item{GOBPTab}{GOBPTab}

\item{GOCCTab}{GOCCTab}

\item{PathTab}{PathTab}

\item{nBar}{nBar}

\item{nRGTab}{nRGTab}
}
\value{
EAbarplot
}
\description{
EAbarplot plots the result from EAcomplete
}
\examples{
Genelist <- c("FN1","COL1A1")
ansEA <- EAcomplete(TFname="DEA genes Normal Vs Tumor",Genelist)
EAbarplot(tf = rownames(ansEA$ResBP),
         GOBPTab = ansEA$ResBP,
         GOCCTab = ansEA$ResCC,
         GOMFTab = ansEA$ResMF,
        PathTab = ansEA$ResPat,
         nRGTab = Genelist,
         nBar = 10)
\dontrun{
Genelist <- rownames(dataDEGsFiltLevel)
system.time(ansEA <- EAcomplete(TFname="DEA genes Normal Vs Tumor",Genelist))
# Enrichment Analysis EA (TCGAVisualize)
# Gene Ontology (GO) and Pathway enrichment barPlot
EAbarplot(tf = rownames(ansEA$ResBP),
         GOBPTab = ansEA$ResBP,
         GOCCTab = ansEA$ResCC,
         GOMFTab = ansEA$ResMF,
        PathTab = ansEA$ResPat,
         nRGTab = Genelist,
         nBar = 10)
}
}

