% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{EB_PCBC_stemSig}
\alias{EB_PCBC_stemSig}
\title{A numeric vector with stem cell (SC)-derived embryoid bodies (EB) signature trained on PCBC's dataset}
\format{
A numeric vector with 12956 genes
}
\description{
A numeric vector with stem cell (SC)-derived embryoid bodies (EB) signature trained on PCBC's dataset
}
\keyword{internal}
