% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{EnrichmentAnalysis}
\alias{EnrichmentAnalysis}
\title{EnrichmentAnalysis}
\usage{
EnrichmentAnalysis(GeneName, RegulonList, TableEnrichment, IPAGenes, GOtype,
  FDRThresh = 0.01)
}
\arguments{
\item{GeneName}{GeneName}

\item{RegulonList}{RegulonList}

\item{TableEnrichment}{TableEnrichment}

\item{GOtype}{GOtype}

\item{FDRThresh}{FDRThresh}
}
\value{
EAcomplete plot
}
\description{
EnrichmentAnalysis
}

