% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{ER_status_BRCA}
\alias{ER_status_BRCA}
\title{ER_status_BRCA}
\usage{
ER_status_BRCA(barcode, ER, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{ER}{ER}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
ER_status_BRCA
}
\description{
ER_status_BRCA
}
\examples{
clin.query <- GDCquery(project = "TCGA-BRCA", data.category = "Clinical",
                      barcode = c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"))
tryCatch(GDCdownload(clin.query), error = function(e) GDCdownload(clin.query, method = "client"))
clinical.patient <- GDCprepare_clinic(clin.query, clinical.info = "patient")
ER_status_BRCA(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),
"Positive",clinical.patient)
}
\keyword{internal}

