% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{gbm.exp.harmonized}
\alias{gbm.exp.harmonized}
\title{A RangedSummarizedExperiment two samples with gene expression data from vignette
aligned against hg38}
\format{A RangedSummarizedExperiment: 56963 genes, 2 samples}
\description{
A RangedSummarizedExperiment two samples with gene expression data from vignette
aligned against hg38
}
\keyword{internal}
