% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{gbm.exp.legacy}
\alias{gbm.exp.legacy}
\title{A RangedSummarizedExperiment two samples with gene expression data from vignette
aligned against hg19}
\format{A RangedSummarizedExperiment: 21022 genes, 2 samples}
\description{
A RangedSummarizedExperiment two samples with gene expression data from vignette
aligned against hg19
}
\keyword{internal}
