% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{GDCdownload}
\alias{GDCdownload}
\title{Download GDC data}
\usage{
GDCdownload(query, token.file, method = "api")
}
\arguments{
\item{query}{A query for GDCquery function}

\item{token.file}{Token file to download controleed data}

\item{method}{Use api method or gdc client tool (API is faster)}
}
\value{
Shows the output from the GDC transfer tools
}
\description{
Uses GDC transfer tool to download gdc data
  The user can use query or manifest.file argument (not both at the same time)
  The data from query will be save in a folder: project/data.category
  The data from manifest.file  will be save in a folder: manifest.file.without.extension/
}

