% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{GDCPrepare_clinic}
\alias{GDCPrepare_clinic}
\title{Parsing clinical xml files}
\usage{
GDCPrepare_clinic(query, clinical.info)
}
\arguments{
\item{query}{Result from GDCquery, with data.category set to Clinical}

\item{clinical.info}{Which information should be retrieved.
Options: drug, admin, follow_up,radiation, patient, stage_event or new_tumor event}
}
\description{
This function receives the query argument and parses the clinical xml files
based on the desired information
}
\examples{
query <- GDCquery(project = "TCGA-COAD",
                  data.category = "Clinical",
                  barcode = c("TCGA-RU-A8FL","TCGA-AA-3972"))
GDCDownload(query)
clinical <- GDCPrepare_clinic(query,"patient")
clinical.drug <- GDCPrepare_clinic(query,"drug")
clinical.radiation <- GDCPrepare_clinic(query,"radiation")
clinical.admin <- GDCPrepare_clinic(query,"admin")
}

