% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAQuery.R
\name{GDCquery}
\alias{GDCquery}
\title{Query GDC data}
\usage{
GDCquery(project, data.category, data.type, workflow.type, legacy = FALSE,
  platform, file.type, barcode, sample.type)
}
\arguments{
\item{project}{A valid project (see list with TCGAbiolinks:::getGDCprojects()$project_id)]}

\item{data.category}{A valid project (see list with TCGAbiolinks:::getProjectSummary(project))}

\item{data.type}{A data type to filter the files to download}

\item{workflow.type}{GDC workflow type}

\item{legacy}{Search in the legacy repository}

\item{platform}{Example:
\tabular{ll}{
CGH- 1x1M_G4447A                   \tab IlluminaGA_RNASeqV2   \cr
AgilentG4502A_07                  \tab IlluminaGA_mRNA_DGE    \cr
Human1MDuo                        \tab HumanMethylation450    \cr
HG-CGH-415K_G4124A                \tab IlluminaGA_miRNASeq    \cr
HumanHap550                       \tab IlluminaHiSeq_miRNASeq \cr
ABI                               \tab H-miRNA_8x15K  \cr
HG-CGH-244A                       \tab SOLiD_DNASeq                \cr
IlluminaDNAMethylation_OMA003_CPI \tab IlluminaGA_DNASeq_automated   \cr
IlluminaDNAMethylation_OMA002_CPI \tab HG-U133_Plus_2                 \cr
HuEx- 1_0-st-v2 \tab Mixed_DNASeq                  \cr
H-miRNA_8x15Kv2 \tab IlluminaGA_DNASeq_curated      \cr
MDA_RPPA_Core   \tab IlluminaHiSeq_TotalRNASeqV2    \cr
HT_HG-U133A     \tab IlluminaHiSeq_DNASeq_automated \cr
diagnostic_images                 \tab microsat_i                     \cr
IlluminaHiSeq_RNASeq              \tab SOLiD_DNASeq_curated           \cr
IlluminaHiSeq_DNASeqC             \tab Mixed_DNASeq_curated           \cr
IlluminaGA_RNASeq                 \tab IlluminaGA_DNASeq_Cont_automated  \cr
IlluminaGA_DNASeq                 \tab IlluminaHiSeq_WGBS             \cr
pathology_reports                 \tab IlluminaHiSeq_DNASeq_Cont_automated\cr
Genome_Wide_SNP_6                 \tab bio                            \cr
tissue_images                     \tab Mixed_DNASeq_automated         \cr
HumanMethylation27                \tab Mixed_DNASeq_Cont_curated      \cr
IlluminaHiSeq_RNASeqV2            \tab Mixed_DNASeq_Cont
}}

\item{file.type}{To be used in the legacy database for some platforms,
to define which file types to be used.}

\item{barcode}{A list of barcodes to filter the files to download}

\item{sample.type}{A sample type to filter the files to download}
}
\value{
A data frame with the results and the parameters used
}
\description{
Uses GDC API to search for search, it searches for both controlled and
  open-acess data.
  For GDC data arguments project, data.category, data.type and workflow.type should be used
  For the legacy data arguments project, data.category, platform and/or file.extension should be used.
}

