% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{GDCquery_maf}
\alias{GDCquery_maf}
\title{Retrieve open access maf files from GDC server}
\usage{
GDCquery_maf(project, pipelines = NULL, save.csv = FALSE,
  directory = "GDCdata")
}
\arguments{
\item{project}{A valid GDC project}

\item{pipelines}{Four separate variant calling pipelines are implemented for GDC data harmonization.
Options: muse, varscan2, somaticsniper, mutect2. For more information:
https://gdc-docs.nci.nih.gov/Data/Bioinformatics_Pipelines/DNA_Seq_Variant_Calling_Pipeline/}

\item{save.csv}{Write maf file into a csv document}

\item{directory}{Directory/Folder where the data will downloaded. Default: GDCdata}
}
\value{
A data frame with the maf file information
}
\description{
GDCquery_maf uses the following guide to download maf files
  https://gdc-docs.nci.nih.gov/Data/Release_Notes/Data_Release_Notes/
}
\examples{
acc.muse.maf <- GDCquery_maf("TCGA-CHOL", pipelines = "muse")
\dontrun{
  acc.varscan2.maf <- GDCquery_maf("TCGA-CHOL", pipelines = "varscan2")
   acc.somaticsniper.maf <- GDCquery_maf("TCGA-CHOL", pipelines = "somaticsniper")
   acc.mutect.maf <- GDCquery_maf("TCGA-CHOL", pipelines = "mutect2")
}
}
