% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{gender_BRCA}
\alias{gender_BRCA}
\title{gender_BRCA}
\usage{
gender_BRCA(barcode, gender, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{gender}{gender}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA
}
\description{
gender_BRCA
}
\examples{
clin.query <- GDCquery(project = "TCGA-BRCA", data.category = "Clinical",
                      barcode = c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"))
tryCatch(GDCdownload(clin.query), error = function(e) GDCdownload(clin.query, method = "client"))
clinical.patient <- GDCprepare_clinic(clin.query, clinical.info = "patient")
gender_BRCA (c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),"FEMALE",clinical.patient)
}
\keyword{internal}

