% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{gender_BRCA_test}
\alias{gender_BRCA_test}
\title{gender_BRCA_test}
\usage{
gender_BRCA_test(barcode, gender, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{gender}{gender}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA_test
}
\description{
gender_BRCA_test
}

