% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{gene.location}
\alias{gene.location}
\title{Genome Reference Consortium Human (hg19) retrived from biomaRt library
The data set contains the following fields:
\itemize{
\item chromosome_name
\item start_position
\item end_position
\item strand
\item external_gene_name
\item entrezgene
}}
\format{A data frame with 25703 rows and 6 variables}
\description{
Genome Reference Consortium Human (hg19) retrived from biomaRt library
The data set contains the following fields:
\itemize{
\item chromosome_name
\item start_position
\item end_position
\item strand
\item external_gene_name
\item entrezgene
}
}
\keyword{internal}

