% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{geneInfo}
\alias{geneInfo}
\title{geneInfo for normalization of RNAseq data}
\format{A data frame with 20531 rows and 2 variables}
\description{
geneInfo for normalization of RNAseq data
}
\keyword{internal}
