% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{GenesCutID}
\alias{GenesCutID}
\title{GenesCutID}
\usage{
GenesCutID(GeneList)
}
\arguments{
\item{GeneList}{GeneList}
}
\value{
list of gene symbol without IDs
}
\description{
GenesCutID
}
\examples{
GenesCutID(c("CRKL|1399","TADA2A|6871","KRT76|51350"))
}

