% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{GeneSplitRegulon}
\alias{GeneSplitRegulon}
\title{GeneSplitRegulon}
\usage{
GeneSplitRegulon(Genelist, Sep)
}
\arguments{
\item{Genelist}{Genelist}

\item{Sep}{Sep}
}
\value{
GeneSplitRegulon
}
\description{
GeneSplitRegulon
}
\keyword{internal}
