% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAVersion.R
\name{get.barcodes}
\alias{get.barcodes}
\title{Get bar code info}
\usage{
get.barcodes(platform, disease, center)
}
\arguments{
\item{magetab.url}{path to mage-tab folder}
}
\description{
Get bar code info using sdrf files
              Obs1: it takes a lot of time, is there a better way
              than downloading sdrf?
              Obs2: Not all folders has mage-tab file
              Solution 1: search for barcode pattern in the filenames
              Solution 2: maf files has it inside. Should use downloaded files.
}
\keyword{internal}

