% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{get.GRCh.bioMart}
\alias{get.GRCh.bioMart}
\title{Get hg19 or hg38 information from biomaRt}
\usage{
get.GRCh.bioMart(genome = "hg19", as.granges = FALSE)
}
\arguments{
\item{genome}{hg38 or hg19}

\item{as.granges}{Output as GRanges or data.frame}
}
\description{
Get hg19 or hg38 information from biomaRt
}
