% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_IDs}
\alias{get_IDs}
\title{Extract information from TCGA barcodes.}
\usage{
get_IDs(data)
}
\arguments{
\item{data}{numeric matrix, each row represents a gene, each column represents a sample}
}
\value{
data frame with columns 'project', 'tss','participant', 'sample', "portion", "plate", "center", "condition"
}
\description{
get_IDs allows user to extract metadata from barcodes. The dataframe returned has columnns for
 'project', 'tss','participant', 'sample', "portion", "plate", and "center"
}
