% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getDataCategorySummary}
\alias{getDataCategorySummary}
\title{Create a Summary table for each sample in a project saying if it contains
or not files for a certain data category}
\usage{
getDataCategorySummary(project)
}
\arguments{
\item{project}{A GDC project}
}
\value{
A data frame
}
\description{
Create a Summary table for each sample in a project saying if it contains
or not files for a certain data category
}
\examples{
summary <- getDataCategorySummary("TCGA-ACC")
}
\author{
Tiago Chedraoui Silva
}
