% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getGDCInfo}
\alias{getGDCInfo}
\title{Check GDC server status}
\usage{
getGDCInfo()
}
\value{
Return true all status
}
\description{
Check GDC server status using the api
  https://gdc-api.nci.nih.gov/status
}
\examples{
info <- getGDCInfo()
}
