% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAInternal.R
\name{getGDCprojects}
\alias{getGDCprojects}
\title{Retrieve all GDC projects}
\usage{
getGDCprojects()
}
\value{
A data frame with last GDC projects
}
\description{
getGDCprojects uses the following api to get projects
  https://gdc-api.nci.nih.gov/projects
}
\examples{
projects <- getGDCprojects()
}
