% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getGistic}
\alias{getGistic}
\title{Download GISTIC data from firehose}
\usage{
getGistic(disease, type = "thresholded")
}
\arguments{
\item{disease}{TCGA disease. Option available in
http://gdac.broadinstitute.org/runs/analyses__latest/data/}

\item{type}{Results type: thresholded or data}
}
\description{
Download GISTIC data from firehose from
http://gdac.broadinstitute.org/runs/analyses__latest/data/
}
