% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getManifest}
\alias{getManifest}
\title{Get a Manifest from GDCquery output that can be used with GDC-client}
\usage{
getManifest(query, save = F)
}
\arguments{
\item{query}{A query for GDCquery function}

\item{save}{Write Manifest to a txt file (tab separated)}
}
\description{
Get a Manifest from GDCquery output that can be used with GDC-client
}
\examples{
query <- GDCquery(project = "TARGET-AML",
                  data.category = "Transcriptome Profiling",
                  data.type = "Gene Expression Quantification",
                  workflow.type = "HTSeq - Counts",
                  barcode = c("TARGET-20-PADZCG-04A-01R","TARGET-20-PARJCR-09A-01R"))
getManifest(query)
}
