% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getMC3MAF}
\alias{getMC3MAF}
\title{Retrieve open access mc3 MAF file from GDC server}
\usage{
getMC3MAF()
}
\value{
A data frame with the MAF file information from https://gdc.cancer.gov/about-data/publications/mc3-2017
}
\description{
Download data from https://gdc.cancer.gov/about-data/publications/mc3-2017
  https://gdc-docs.nci.nih.gov/Data/Release_Notes/Data_Release_Notes/
}
\examples{
\dontrun{
   maf <- getMC3MAF()
}
}
