% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getNbCases}
\alias{getNbCases}
\title{Get Number of cases in GDC for a project}
\usage{
getNbCases(project, data.category, legacy = FALSE)
}
\arguments{
\item{project}{A  GDC project}

\item{data.category}{A  GDC project data category}

\item{legacy}{Select between Harmonized or Legacy database}
}
\description{
Get Number of cases in GDC for a project
}
\examples{
\dontrun{
getNbCases("TCGA-ACC","Clinical")
getNbCases("CPTAC-2","Clinical")
}
}
\author{
Tiago Chedraoui Silva
}
