% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getProjectSummary}
\alias{getProjectSummary}
\title{Get Project Summary from GDC}
\usage{
getProjectSummary(project, legacy = FALSE)
}
\arguments{
\item{project}{A  GDC project}

\item{legacy}{Select between Harmonized or Legacy database}
}
\description{
Get Project Summary from GDC
}
\examples{
\dontrun{
getProjectSummary("TCGA-ACC")
getProjectSummary("CPTAC-2")
}
}
