% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getSampleFilesSummary}
\alias{getSampleFilesSummary}
\title{Retrieve summary of files per sample in a project}
\usage{
getSampleFilesSummary(project, files.access = NA)
}
\arguments{
\item{project}{A GDC project}

\item{files.access}{Filter by file access ("open" or "controlled").
Default: no filter}
}
\value{
A data frame with the maf file information
}
\description{
Retrieve the numner of files under each
  data_category + data_type + experimental_strategy + platform
  Almost like https://portal.gdc.cancer.gov/exploration
}
\examples{
   summary <- getSampleFilesSummary("TCGA-UCS")
\dontrun{
   summary <- getSampleFilesSummary(c("TCGA-OV","TCGA-ACC"))
}
}
\author{
Tiago Chedraoui Silva
}
