% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{HER_status_BRCA_test}
\alias{HER_status_BRCA_test}
\title{HER_status_BRCA_test}
\usage{
HER_status_BRCA_test(barcode, HER, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{HER}{HER}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
HER_status_BRCA_test
}
\description{
HER_status_BRCA_test
}

