% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{hg19genes}
\alias{hg19genes}
\title{hg19genes data frame from Human (hg19) genome indicating different
characteristics of genes e.g. chr, region, start, end, gene_name}
\format{A data frame with 869204 rows and 5 variables}
\description{
hg19genes data frame from Human (hg19) genome indicating different
characteristics of genes e.g. chr, region, start, end, gene_name
}
\keyword{internal}

