% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{IPAbarplot}
\alias{IPAbarplot}
\title{IPAbarplot}
\usage{
IPAbarplot(tf, GOMFTab, GOBPTab, GOCCTab, PathTab, nBar, nRGTab)
}
\arguments{
\item{tf}{tf}

\item{GOMFTab}{GOMFTab}

\item{GOBPTab}{GOBPTab}

\item{GOCCTab}{GOCCTab}

\item{PathTab}{PathTab}

\item{nBar}{nBar}

\item{nRGTab}{nRGTab}
}
\value{
GeneSplitRegulon
}
\description{
IPAbarplot
}

