% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{isServeOK}
\alias{isServeOK}
\title{Check GDC server status is OK}
\usage{
isServeOK()
}
\value{
Return true if status is ok
}
\description{
Check GDC server status using the api
  https://api.gdc.cancer.gov/status
}
\examples{
status <- isServeOK()
}
