\name{luadtools-package}
\alias{luadtools-package}
\alias{luadtools}
\title{luadtools: TCGA-LUAD RNA-seq analysis helpers}
\description{Reusable helpers to download TCGA-LUAD RNA-seq counts, run DESeq2 tumor-vs-normal analyses, and produce standard plots and KEGG enrichment outputs.}
\docType{package}
\keyword{package}
