% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{MatchedCoupledSampleTypes}
\alias{MatchedCoupledSampleTypes}
\title{MatchedCoupledSampleTypes}
\usage{
MatchedCoupledSampleTypes(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
MultiSampleTypes
}
\description{
MatchedCoupledSampleTypes
}

