% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{MatchedCoupledSampleTypes_test}
\alias{MatchedCoupledSampleTypes_test}
\title{MatchedCoupledSampleTypes_test}
\usage{
MatchedCoupledSampleTypes_test(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
MultiSampleTypes_test
}
\description{
MatchedCoupledSampleTypes_test
}

