% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{matchedMetExp}
\alias{matchedMetExp}
\title{Get GDC primary tumors samples with both DNA methylation (HM450K) and Gene expression data}
\usage{
matchedMetExp(project, n = NULL)
}
\arguments{
\item{project}{A GDC project}

\item{n}{Number of samples to return. If NULL return all (default)}
}
\value{
A vector of barcodes
}
\description{
For a given TCGA project it gets the  primary tumors samples (barcode) with both
DNA methylation and Gene expression data from GDC database
}
\examples{
# Get ACC samples with both  DNA methylation (HM450K) and gene expression aligned to hg19
samples <- matchedMetExp("TCGA-UCS")
}
