% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{MESO_PCBC_stemSig}
\alias{MESO_PCBC_stemSig}
\title{A numeric vector with SC-derived mesoderm (MESO) signature trained on PCBC's dataset}
\format{
A numeric vector with 12956 genes
}
\description{
A numeric vector with SC-derived mesoderm (MESO) signature trained on PCBC's dataset
}
\keyword{internal}
