% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{met.gbm.27k}
\alias{met.gbm.27k}
\title{A DNA methylation RangedSummarizedExperiment for 8 samples (only first 20 probes)
aligned against hg19}
\format{A RangedSummarizedExperiment: 20 probes, 8 samples}
\description{
A DNA methylation RangedSummarizedExperiment for 8 samples (only first 20 probes)
aligned against hg19
}
\keyword{internal}
