% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{MultiSampleTypes}
\alias{MultiSampleTypes}
\title{MultiSampleTypes}
\usage{
MultiSampleTypes(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
MultiSampleTypes
}
\description{
MultiSampleTypes
}
\examples{
# selection of normal samples "NT"
dataFilt <- "TCGA-06-0125-01A-01D-A45W-05"
samplesNT <- MultiSampleTypes(dataFilt, typesample = c("NT"))
}

