% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{MultiSampleTypes_test}
\alias{MultiSampleTypes_test}
\title{MultiSampleTypes_test}
\usage{
MultiSampleTypes_test(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
MultiSampleTypes_test
}
\description{
MultiSampleTypes_test
}

