% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{organizeMethylationDataFrame}
\alias{organizeMethylationDataFrame}
\title{organize TCGA Methylation Data}
\usage{
organizeMethylationDataFrame(wd = getwd())
}
\arguments{
\item{wd}{Directory with the files}
}
\value{
Methylation betavalues table
}
\description{
Organize TCGA methylation data for the mean methylation analysis.
}
\details{
Input: a directory with DNA methylation data,
   Output: a data frame with DNA methylation values
   where rows are the probes names and columns are paciente ID
   Execution: read all files inside the directory and merge it by
   probes (Composite.Element.REF)
}
\examples{
met <- organizeMethylationDataFrame(wd = "data")
}

