% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{organizeMethylationMetaDataFrame}
\alias{organizeMethylationMetaDataFrame}
\title{organize TCGA Methylation MetaData}
\usage{
organizeMethylationMetaDataFrame(wd = NULL)
}
\arguments{
\item{wd}{Directory with the files}
}
\value{
\code{invisible (metadata)}
}
\description{
Organize TCGA methylation metadata for the mean methylation analysis.
}
\examples{
\dontrun{
met.md <- organizeMethylationMetaDataFrame("data")
}
}

