% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{pancan2018}
\alias{pancan2018}
\title{A data frame with all TCGA molecular subtypes}
\format{A data frame with 7,734 lines and 10 columns}
\description{
A data frame with all TCGA molecular subtypes
}
\keyword{internal}
