% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{PanCan2018_subtypes}
\alias{PanCan2018_subtypes}
\title{Retrieve table with TCGA molecular subtypes}
\usage{
PanCan2018_subtypes()
}
\value{
a data.frame with barcode and molecular subtypes for 24 cancer types
}
\description{
PanCan2018_subtypes is a curated table with molecular subtypes for 24 TCGA cancer types
}
\examples{
molecular.subtypes <- PanCan2018_subtypes()
}
