% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{PanCan33_subtypes}
\alias{PanCan33_subtypes}
\title{Retrieve table with TCGA molecular subtypes}
\usage{
PanCan33_subtypes()
}
\value{
a data.frame with barcode and molecular subtypes for 24 cancer types
}
\description{
PanCan33_subtypes is a curated table with molecular subtypes for 24 TCGA cancer types
}
\examples{
molecular.subtypes <- PanCan33_subtypes()
}
