% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{plotPCAforGroups}
\alias{plotPCAforGroups}
\title{plotPCAforGroups}
\usage{
plotPCAforGroups(dataFilt, dataDEGsFiltLevel, ntopgenes)
}
\arguments{
\item{dataFilt}{dataFilt}

\item{dataDEGsFiltLevel}{dataDEGsFiltLevel}

\item{ntopgenes}{ntopgenes}
}
\value{
PCA plot
}
\description{
plotPCAforGroups
}

