% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{PR_status_BRCA_test}
\alias{PR_status_BRCA_test}
\title{PR_status_BRCA_test}
\usage{
PR_status_BRCA_test(barcode, PR, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{PR}{PR}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
PR_status_BRCA_test
}
\description{
PR_status_BRCA_test
}

