% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{RnaSeqFilt}
\alias{RnaSeqFilt}
\title{Filtering rnaseq by quantile}
\usage{
RnaSeqFilt(TableRnaseq, QuantileThresh)
}
\arguments{
\item{TableRnaseq}{TableRnaseq}

\item{QuantileThresh}{QuantileThresh}
}
\value{
table filtered
}
\description{
Filtering rnaseq by quantile
}
\examples{
\dontrun{
dataNorm <- TCGAbiolinks::RnaSeqNormalization(dataBRCA, geneInfo)
dataFilt <- RnaSeqFilt(dataNorm, 0.25)
}
}

