% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{RnaSeqNormalization}
\alias{RnaSeqNormalization}
\title{RnaSeqNormalization}
\usage{
RnaSeqNormalization(TCGA_RnaseqTable, geneInfo)
}
\arguments{
\item{TCGA_RnaseqTable}{Rnaseq numeric matrix, each row represents a gene,
each column represents a sample}

\item{geneInfo}{Information matrix of 20531 genes about geneLength and gcContent}
}
\value{
Rnaseq matrix normalized with counts slot holds the count data as a matrix
of non-negative integer count values, one row for each observational unit (gene or the like),
and one column for each sample.
}
\description{
RnaSeqNormalization allows user to normalize mRNA transcripts and miRNA,
   using EDASeq package. Normalization for RNA-Seq Numerical and graphical summaries of RNA-Seq read data. Within-lane normalization procedures
   to adjust for GC-content effect (or other gene-level effects) on read counts:
   loess robust local regression, global-scaling, and full-quantile normalization
   (Risso et al., 2011). Between-lane normalization procedures to adjust for
   distributional differences between lanes (e.g., sequencing depth): global-scaling and full-quantile normalization (Bullard et al., 2010).
   For istance returns all mRNA or miRNA with mean across all
   samples, higher than the threshold defined quantile mean across all samples.
   RnaSeqNormalization performs normalization using following functions from EDASeq
   1. EDASeq::newSeqExpressionSet
   2. EDASeq::withinLaneNormalization
   3. EDASeq::betweenLaneNormalization
   4. EDASeq::counts
}
\examples{
dataNorm <- TCGAbiolinks::RnaSeqNormalization(dataBRCA, geneInfo)
}

