% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAintegrate.R
\name{samplesfilter}
\alias{samplesfilter}
\title{Filtering sample output from TCGAQuery}
\usage{
samplesfilter(query)
}
\arguments{
\item{query}{metaData output from TCGAQuery}
}
\value{
list of samples for a tumor
}
\description{
Filtering sample output from TCGAQuery
}

