% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{SampleTypes_test}
\alias{SampleTypes_test}
\title{SampleTypes_test}
\usage{
SampleTypes_test(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
SampleTypes_test
}
\description{
SampleTypes_test
}

