% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{splitAPICall}
\alias{splitAPICall}
\title{internal function to break a huge API call into smaller ones
so it repects the max character limit of a string}
\usage{
splitAPICall(FUN, step = 20, items)
}
\arguments{
\item{FUN}{function that calls the API}

\item{step}{How many items to be evaluated per API call}

\item{items}{vector of items to be using within the function
(list of barcodes, aliquot ids, etc)}
}
\description{
internal function to break a huge API call into smaller ones
so it repects the max character limit of a string
}
\author{
Tiago Chedraoui Silva
}
