% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{stage_BRCA}
\alias{stage_BRCA}
\title{stage_BRCA}
\usage{
stage_BRCA(barcode, stage, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{stage}{stage}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA
}
\description{
stage_BRCA
}
\examples{
clin.query <- GDCquery(project = "TCGA-BRCA", data.category = "Clinical",
                      barcode = c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"))
tryCatch(GDCdownload(clin.query), error = function(e) GDCdownload(clin.query, method = "client"))
clinical.patient <- GDCprepare_clinic(clin.query, clinical.info = "patient")
stage_BRCA(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),"stage_IX",clinical.patient)
}
\keyword{internal}

