% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{stage_BRCA_test}
\alias{stage_BRCA_test}
\title{stage_BRCA_test}
\usage{
stage_BRCA_test(barcode, stage, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{stage}{stage}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA_test
}
\description{
stage_BRCA_test
}

