% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{starbursAnalysis}
\alias{starbursAnalysis}
\title{Preparing methylation data and expression data}
\usage{
starbursAnalysis(met, expression)
}
\arguments{
\item{met}{methylation data}

\item{expression}{expression data}
}
\value{
Dataframe with methylation and expression data
}
\description{
Preparing methylation data and expression data
}

