% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{starburstPlot}
\alias{starburstPlot}
\title{Create starburst plot}
\usage{
starburstPlot(met, exp, filename = "volcano.pdf",
  ylab = paste0("Gene Expression\\nlog10 of the",
  "adjusted Significance (FDR)"),
  xlab = paste0("DNA Methylation\\nlog10 of the",
  " adjusted Significance (FDR)"), title = "Starburst Plot",
  legend = "Methylation/Expression Relation", color = c(`1` = "black", `2` =
  "purple", `3` = "darkgreen", `4` = "blue", `5` = "darkred", `6` = "red", `7` =
  "green", `8` = "yellow", `9` = "orange"), label = c(`1` = "Not Significant",
  `2` = "Up & Hypo", `3` = "Down & Hypo", `4` = "hypo", `5` = "hyper", `6` =
  "Up", `7` = "Down", `8` = "Up & Hyper", `9` = "Down & Hyper"), xlim = NULL,
  ylim = NULL, p.cut = 0.05)
}
\arguments{
\item{met}{SummarizedExperiment with methylation data obtained from the
TCGAPrepare. Expected colData columns: diffmean,  p.value.adj  and p.value
Execute volcanoPlot function in order to obtain these values for the object.}

\item{exp}{SummarizedExperiment with methylation data obtained from the
TCGAPrepare. Expected colData columns: diffmean,  p.value.adj  and p.value}

\item{filename}{pdf filename}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title}

\item{legend}{legend title}

\item{color}{vector of colors to be used in graph}

\item{label}{vector of labels to be used in graph}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}

\item{p.cut}{p value cut-off}
}
\value{
Save a starburst plot
}
\description{
Create starburst plot
}
\details{
Input: data with gene expression/methylation expression
   Output: starburst plot
}
\examples{
starburstPlot(gene.met)
}

