% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{survivalAnalysis}
\alias{survivalAnalysis}
\title{Creates survival analysis}
\usage{
survivalAnalysis(data, clusterCol = NULL, legend = "Legend", cutoff = 0,
  main = "Kaplan-Meier Overall Survival Curves",
  ylab = "PROBABILITY OF SURVIVAL", xlab = "TIME SINCE DIAGNOSIS (DAYS)",
  filename = "survival.pdf", color = c("green", "firebrick4", "orange3",
  "blue"))
}
\arguments{
\item{data}{TCGA Clinical patient with the information days_to_death}

\item{clusterCol}{Column with groups to plot. This is a mandatory field, the
caption will be based in this column}

\item{legend}{Legend title of the figure}

\item{cutoff}{xlim This parameter will be a limit in the x-axis. That means, that
patients with days_to_deth > cutoff will be set to Alive.}

\item{main}{main title of the plot}

\item{ylab}{y axis text of the plot}

\item{xlab}{x axis text of the plot}

\item{filename}{The name of the pdf file}

\item{color}{Define the colors of the lines.}
}
\value{
Survival plot
}
\description{
Creates a survival plot from TCGA patient clinical data
using survival library. It uses the fields days_to_death and vital, plus a
columns for groups.
}
\examples{
death_days_to <- floor(runif(200, 1, 1000))
vital_status <- c(rep("Dead",200))
groups <- c(rep(c("G1","G2"),c(100,100)))
df <- data.frame(death_days_to,vital_status,groups)
survivalAnalysis(df,clusterCol="groups")
\dontrun{
clinical <- clinic("gbm","clinical_patient")
survivalAnalysis(clinical,"gender", filename = "surv.pdf", legend="Gender")
}
}

