% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SurvivalCoxNET.R
\name{SurvivalCoxNET}
\alias{SurvivalCoxNET}
\title{SurvivalCoxNET}
\usage{
SurvivalCoxNET(clinical_patient, dataGE, Genelist, scoreConfidence = 700,
  titlePlot = "SurvivalCoxNET Example")
}
\arguments{
\item{clinical_patient}{clinical_patient}

\item{dataGE}{dataGE}

\item{Genelist}{Genelist}

\item{scoreConfidence}{restrict to those edges with high confidence (eg. score>=700)}

\item{titlePlot}{titlePlot}
}
\value{
net IGRAPH with attr: name (v/c), seqid (v/c), geneid (v/n), symbol (v/c), description (v/c) ...
}
\description{
Survival analysis with univariate Cox regression package (dnet)
}

