% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{tabDEGsTFPubmed}
\alias{tabDEGsTFPubmed}
\title{TCGA data matrix BRCA DEGs Pubmed}
\format{A data frame with 10 rows and 8 variables}
\description{
TCGA data matrix BRCA DEGs Pubmed
}
\keyword{internal}

