% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pam50Info.R
\docType{data}
\name{TabSubtypesCol_merged}
\alias{TabSubtypesCol_merged}
\title{TCGA samples with their Pam50 subtypes}
\format{A data frame with 4768 rows and 3 variables:
\describe{
  \item{samples}{Sample ID from TCGA barcodes, character string}
  \item{subtype}{Pam50 classification, character string}
  \item{color}{color, character string}
  ...
}}
\usage{
TabSubtypesCol_merged
}
\description{
A dataset containing the Sample Ids from TCGA and PAM50 subtyping attributes of 4768 tumor patients
}
\keyword{datasets}
