% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{tabSurvKMcompleteDEGs}
\alias{tabSurvKMcompleteDEGs}
\title{tabSurvKMcompleteDEGs}
\format{A data frame with 200 rows and 7 variables}
\description{
tabSurvKMcompleteDEGs
}
\keyword{internal}

