% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{tcga.db}
\alias{tcga.db}
\title{TCGA metadata database.
The data set contains the following fields:
\itemize{
\item addedDate Date when sample was added to database
\item baseName name of the sample folder
\item deployLocation Path of the sample folder
\item barcode list of barcode
\item id TCGA id
\item isLatest Is the lasted version of the sample?
\item name Sample name
\item revision Sample revision
\item serialIndex Sample serial Index
\item Center Center name
\item Platform Sample platform
\item Disease Sample disease
}}
\format{A data frame with 4738 rows and 12 variables}
\description{
TCGA metadata database.
The data set contains the following fields:
\itemize{
\item addedDate Date when sample was added to database
\item baseName name of the sample folder
\item deployLocation Path of the sample folder
\item barcode list of barcode
\item id TCGA id
\item isLatest Is the lasted version of the sample?
\item name Sample name
\item revision Sample revision
\item serialIndex Sample serial Index
\item Center Center name
\item Platform Sample platform
\item Disease Sample disease
}
}
\keyword{internal}

