% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{TCGA_MolecularSubtype}
\alias{TCGA_MolecularSubtype}
\title{Retrieve molecular subtypes for given TCGA barcodes}
\usage{
TCGA_MolecularSubtype(barcodes)
}
\arguments{
\item{barcodes}{is a vector of TCGA barcodes}
}
\value{
List with $subtypes attribute as a dataframe with barcodes, samples, subtypes, and colors. The $filtered attribute is returned as filtered samples with no subtype info
}
\description{
TCGA_MolecularSubtype Retrieve molecular subtypes from TCGA consortium for a given set of barcodes
}
\examples{
TCGA_MolecularSubtype("TCGA-60-2721-01A-01R-0851-07")
}
