% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAanalyze.R
\name{TCGAanalyze_analyseGRN}
\alias{TCGAanalyze_analyseGRN}
\title{Generate network}
\usage{
TCGAanalyze_analyseGRN(TFs, normCounts, kNum)
}
\arguments{
\item{TFs}{a vector of genes.}

\item{normCounts}{is a matrix of gene expression with genes in rows and samples in columns.}

\item{kNum}{the number of nearest neighbors to consider to estimate the mutual information.
Must be less than the number of columns of normCounts.}
}
\value{
an adjacent matrix
}
\description{
TCGAanalyze_analyseGRN perform gene regulatory network.
}
