% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAanalyze.R
\name{TCGAanalyze_Clustering}
\alias{TCGAanalyze_Clustering}
\title{Hierarchical cluster analysis}
\usage{
TCGAanalyze_Clustering(tabDF, method, methodHC = "ward.D2")
}
\arguments{
\item{tabDF}{is a dataframe or numeric matrix, each row represents a gene,
each column represents a sample come from TCGAPrepare.}

\item{method}{is method to be used for generic cluster such as 'hclust' or 'consensus'}

\item{methodHC}{is method to be used for Hierarchical cluster.}
}
\value{
object of class hclust if method selected is 'hclust'.
If method selected is 'Consensus' returns a list of length maxK
(maximum cluster number to evaluate.). Each element is a list containing
consensusMatrix (numerical matrix), consensusTree (hclust), consensusClass
(consensus class asssignments). ConsensusClusterPlus also produces images.
}
\description{
Hierarchical cluster analysis using several methods such as
ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC).
}

