% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAanalyze.R
\name{TCGAanalyze_copynumber}
\alias{TCGAanalyze_copynumber}
\title{Identifying Segment Gain Or Loss (SGOL)}
\usage{
TCGAanalyze_copynumber(data = NULL, threshold = 0)
}
\arguments{
\item{data}{Copy number data from TCGAPrepare}

\item{threshold}{Gain/Loss threshold}
}
\value{
List of list with tables in 2 by 2 comparison
of the top-ranked genes from a linear model fitted by DEA's limma
}
\description{
Identifying Segment Gain Or Loss (SGOL) using findMCR package
Example:
query <- TCGAquery("acc","genome_wide_snp_6", level = 3)
TCGAdownload(query,"trash")
data <- TCGAprepare(query,"trash",  type = "nocnv_hg19")
}
\examples{
\dontrun{
 query <- TCGAquery("acc", platform = "genome_wide_snp_6", level = 3)
 TCGAdownload(query,path = "trash", samples = "TCGA-OR-A5JH-01A-11D-A309-01")
 segment_data <- TCGAprepare(query, dir = "trash",  type = "nocnv_hg19",
                    samples = "TCGA-OR-A5JH-01A-11D-A309-01")
 TCGAanalyze_copynumber(segment_data)
}
}

