% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{TCGAanalyze_DEA_Affy}
\alias{TCGAanalyze_DEA_Affy}
\title{Differentially expression analysis (DEA) using limma package.}
\usage{
TCGAanalyze_DEA_Affy(AffySet, FC.cut = 0.01)
}
\arguments{
\item{AffySet}{A matrix-like data object containing log-ratios or log-expression values
for a series of arrays, with rows corresponding to genes and columns to samples}

\item{FC.cut}{write}
}
\value{
List of list with tables in 2 by 2 comparison
of the top-ranked genes from a linear model fitted by DEA's limma
}
\description{
Differentially expression analysis (DEA) using limma package.
}
\examples{
\dontrun{
to add example
}
}
