% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAanalyze.R
\name{TCGAanalyze_networkInference}
\alias{TCGAanalyze_networkInference}
\title{infer gene regulatory networks}
\usage{
TCGAanalyze_networkInference(data, optionMethod = "clr")
}
\arguments{
\item{data}{expression data, genes in columns, samples in rows}

\item{optionMethod}{inference method, chose from aracne, c3net, clr and mrnet}
}
\value{
an adjacent matrix
}
\description{
TCGAanalyze_networkInference taking expression data as input, this will return an adjacency matrix of interactions
}
