% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{TCGAanalyze_Pathview}
\alias{TCGAanalyze_Pathview}
\title{Generate pathview graph}
\usage{
TCGAanalyze_Pathview(dataDEGs, pathwayKEGG = "hsa05200")
}
\arguments{
\item{dataDEGs}{dataDEGs}

\item{pathwayKEGG}{pathwayKEGG}
}
\value{
an adjacent matrix
}
\description{
TCGAanalyze_Pathview pathway based data integration and visualization.
}
\examples{
\dontrun{
  dataDEGs <- data.frame(mRNA = c("TP53","TP63","TP73"), logFC = c(1,2,3))
  TCGAanalyze_Pathview(dataDEGs)
}
}
