% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{TCGAanalyze_survival}
\alias{TCGAanalyze_survival}
\title{Creates survival analysis}
\usage{
TCGAanalyze_survival(data, clusterCol = NULL, legend = "Legend",
  labels = NULL, cutoff = 0,
  main = "Kaplan-Meier Overall Survival Curves",
  ylab = "Probability of survival", xlab = "Time since diagnosis (days)",
  filename = "survival.pdf", color = NULL, height = 8, width = 12,
  dpi = 300, legend.position = "inside", legend.title.position = "top",
  legend.ncols = 1, add.legend = TRUE, print.value = TRUE,
  add.points = TRUE)
}
\arguments{
\item{data}{TCGA Clinical patient with the information days_to_death}

\item{clusterCol}{Column with groups to plot. This is a mandatory field, the
caption will be based in this column}

\item{legend}{Legend title of the figure}

\item{labels}{labels of the plot}

\item{cutoff}{xlim This parameter will be a limit in the x-axis. That means, that
patients with days_to_deth > cutoff will be set to Alive.}

\item{main}{main title of the plot}

\item{ylab}{y axis text of the plot}

\item{xlab}{x axis text of the plot}

\item{filename}{The name of the pdf file.}

\item{color}{Define the colors of the lines.}

\item{height}{Image height}

\item{width}{Image width}

\item{dpi}{Figure quality}

\item{legend.position}{Legend position ("top", "right","left","bottom")}

\item{legend.title.position}{Legend title position ("top", "right","left","bottom")}

\item{legend.ncols}{Number of columns of the legend}

\item{add.legend}{If true, legend is created. Otherwise names will
be added to the last point in the lines.}

\item{print.value}{Print pvalue in the plot? Default: TRUE}

\item{add.points}{If true, shows each death at the line of survival curves}
}
\value{
Survival plot
}
\description{
Creates a survival plot from TCGA patient clinical data
using survival library. It uses the fields days_to_death and vital, plus a
columns for groups.
}
\examples{
clin <- GDCquery_clinic("TCGA-LGG", type = "clinical", save.csv = FALSE)
TCGAanalyze_survival(clin, clusterCol="gender")
}

