% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAinvestigate.R
\name{TCGAinvestigate}
\alias{TCGAinvestigate}
\title{TCGAinvestigate}
\usage{
TCGAinvestigate(tumor, dataDEGsFiltLevelTF, topgenes)
}
\arguments{
\item{tumor}{tumor}

\item{dataDEGsFiltLevelTF}{dataDEGsFiltLevelTF}

\item{topgenes}{topgenes}
}
\value{
table with number of pubmed related to tfs.
}
\description{
Find most studied TF in pubmed related to a specific cancer,
 disease, or tissue
}
\examples{
\dontrun{
TFs <- EAGenes[EAGenes$Family =="transcription regulator",]
TFs_inDEGs <- intersect(TFs$Gene, dataDEGsFiltLevel$mRNA )
dataDEGsFiltLevelTFs <- dataDEGsFiltLevel[TFs_inDEGs,]
dataDEGsFiltLevelTFs <- dataDEGsFiltLevelTFs[order(dataDEGsFiltLevelTFs$Delta,decreasing = TRUE),]
# Find Pubmed of TF studied related to cancer
tabDEGsTFPubmed <- TCGAinvestigate("breast", dataDEGsFiltLevelTFs, topgenes = 1)
}
}

