% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{TCGAPam50}
\alias{TCGAPam50}
\title{Retrieve PAM50 molecular subtypes for given TCGA barcodes}
\usage{
TCGAPam50(barcodes)
}
\arguments{
\item{barcodes}{is a vector of TCGA barcodes}
}
\value{
List with $subtypes attribute as a dataframe with barcodes, samples, subtypes, and colors. The $filtered attribute is returned as filtered samples with no subtype info
}
\description{
TCGAPam50 Retrieve PAM50 molecular subtypes from TCGA consortium for a given set of barcodes
}
\examples{
Purity.BRCA<-TCGAtumor_purity(barcodes, 0, 0, 0, 0, 0.7)
}
