% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAprepare}
\alias{TCGAprepare}
\title{Read the data from level 3 the experiments and prepare it
 for downstream analysis into a SummarizedExperiment object.}
\usage{
TCGAprepare(query = NULL, dir = NULL, samples = NULL, type = NULL,
  save = FALSE, filename = NULL, summarizedExperiment = TRUE)
}
\arguments{
\item{query}{TCGAquery output}

\item{dir}{Directory with the files downloaded by TCGAdownload}

\item{samples}{List of samples to prepare the data}

\item{type}{Filter the files to prepare.}

\item{save}{Save a rda object with the prepared object?
Default: \code{FALSE}}

\item{filename}{Name of the saved file}

\item{summarizedExperiment}{Output as SummarizedExperiment?
Default: \code{FALSE}}
}
\value{
A SummarizedExperiment object (If SummarizedExperiment = \code{FALSE},
a data.frame)
}
\description{
This function has been replaced by GDCPrepare

List of accepted platforms:
\itemize{
\item AgilentG4502A_07_1/AgilentG4502A_07_2/AgilentG4502A_07_3
\item Genome_Wide_SNP_6
\item H-miRNA_8x15K/H-miRNA_8x15Kv2
\item HG-U133_Plus_2
\item HT_HG-U133A
\item HumanMethylation27
\item HumanMethylation450
\item IlluminaDNAMethylation_OMA002_CPI
\item IlluminaDNAMethylation_OMA003_CPI
\item IlluminaGA_RNASeq
\item IlluminaGA_RNASeqV2
\item IlluminaHiSeq_RNASeq
\item IlluminaHiSeq_RNASeqV2
\item IlluminaHiSeq_TotalRNASeqV2
}
 \strong{Return}The default output is a SummarizedExperiment object.
}
\examples{
\dontrun{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAquery(tumor = "GBM",samples = sample, level = 3)
TCGAdownload(query,path = "exampleData",samples = sample)
data <- TCGAprepare(query, dir="exampleData")
}
}

