% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{TCGAprepare_Affy}
\alias{TCGAprepare_Affy}
\title{Prepare CEL files into an AffyBatch.}
\usage{
TCGAprepare_Affy(ClinData, PathFolder, TabCel)
}
\arguments{
\item{ClinData}{write}

\item{PathFolder}{write}

\item{TabCel}{write}
}
\value{
Normalizd Expression data from Affy eSets
}
\description{
Prepare CEL files into an AffyBatch.
}
\examples{
\dontrun{
to add example
}
}
