% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAprepare_elmer}
\alias{TCGAprepare_elmer}
\title{Prepare the data for ELEMR package}
\usage{
TCGAprepare_elmer(df, platform = NULL, met.na.cut = 0.2, save = FALSE)
}
\arguments{
\item{df}{data frame from TCGAPrepare setting summarizedExperiment to FALSE}

\item{platform}{platform of the data}

\item{na.cut}{Define the percentage of NA that the line should have to
remove the probes}
}
\value{
Matrix prepared for fetch.mee function
}
\description{
Prepare the data for ELEMR package
}
\examples{
df <- data.frame(runif(200, 1e5, 1e6),runif(200, 1e5, 1e6))
rownames(df) <- sprintf("?|\%03d", 1:200)
TCGAprepare_elmer(df,platform="IlluminaHiSeq_RNASeqV2")
}

