% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGA_biolinks_functions.R
\name{TCGAQuery}
\alias{TCGAQuery}
\title{TCGA Query}
\usage{
TCGAQuery(Tumor, siteTCGA)
}
\arguments{
\item{Tumor}{-  tumor code}

\item{siteTCGA}{- Plataform}
}
\description{
TCGA Query
}

