% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{TCGAquery_clinicFilt}
\alias{TCGAquery_clinicFilt}
\title{Filter samples using clinical data}
\usage{
TCGAquery_clinicFilt(barcode, clinical_patient_data, HER = NULL, ER = NULL,
  gender = NULL, PR = NULL, stage = NULL)
}
\arguments{
\item{barcode}{List of barcodes}

\item{clinical_patient_data}{clinical_patient_data obtained with clinic function}

\item{HER}{her2 neu immunohistochemistry receptor status: "Positive" or "Negative"}

\item{ER}{Estrogen receptor status: "Positive" or "Negative"}

\item{gender}{"MALE" or "FEMALE"}

\item{PR}{Progesterone receptor status: "Positive" or "Negative"}

\item{stage}{Pathologic Stage: "stage_IX", "stage_I", "stage_IA", "stage_IB", "stage_IIX",
"stage_IIA", "stage_IIB", "stage_IIIX","stage_IIIA", "stage_IIIB",
"stage_IIIC", "stage_IV" -}
}
\value{
List of samples that matches the filters
}
\description{
This function will return the samples that matches all filters.
  Filters available: HER, ER,gender,PR, stage.
}
\examples{
clin.query <- GDCquery(project = "TCGA-BRCA", data.category = "Clinical",
                      barcode = c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"))
tryCatch(GDCdownload(clin.query), error = function(e) GDCdownload(clin.query, method = "client"))
clinical.patient <- GDCprepare_clinic(clin.query, clinical.info = "patient")
TCGAquery_clinicFilt(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),clinical.patient,
                     HER="Positive", gender="FEMALE",ER="Positive")
}

