% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_integrate.R
\name{TCGAquery_integrate}
\alias{TCGAquery_integrate}
\title{Filtering common samples among platforms from TCGAquery for the same tumor}
\usage{
TCGAquery_integrate(query)
}
\arguments{
\item{query}{is the output of TCGAquery}
}
\value{
table with common samples among platforms from TCGAquery
}
\description{
In order to help the user to have an overview of the number of
samples in commun we created the function `TCGAquery_integrate` that will receive the
data frame returned from `TCGAquery` and produce a matrix n platforms x n platforms
with the values of samples in commum.
}
\examples{
query <- TCGAquery(tumor = 'brca',level = 3)
matSamples <- TCGAquery_integrate(query)
}

