% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAQuery.R
\name{TCGAquery_maf}
\alias{TCGAquery_maf}
\title{Get last maf file for the tumor}
\usage{
TCGAquery_maf(tumor = NULL, center = NULL)
}
\arguments{
\item{tumor}{tumor type}
}
\value{
list of samples for a tumor
}
\description{
Filtering sample output from TCGAquery
}

