% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{TCGAquery_MatchedCoupledSampleTypes}
\alias{TCGAquery_MatchedCoupledSampleTypes}
\title{Retrieve multiple tissue types from the same patients.}
\usage{
TCGAquery_MatchedCoupledSampleTypes(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
a list of samples / barcode filtered by type sample selected
}
\description{
TCGAquery_MatchedCoupledSampleTypes
}
\examples{
 TCGAquery_MatchedCoupledSampleTypes(c("TCGA-B0-4698-01Z-00-DX1",
                             "TCGA-B0-4698-02Z-00-DX1"),
                             c("TP","TR"))
 barcode <- c("TARGET-20-PANSBH-02A-02D","TARGET-20-PANSBH-01A-02D",
              "TCGA-B0-4698-01Z-00-DX1","TCGA-CZ-4863-02Z-00-DX1",
              "TARGET-20-PANSZZ-02A-02D","TARGET-20-PANSZZ-11A-02D",
              "TCGA-B0-4699-01Z-00-DX1","TCGA-B0-4699-02Z-00-DX1"
              )
  TCGAquery_MatchedCoupledSampleTypes(barcode,c("TR","TP"))
}
