% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{TCGAquery_MultiSampleTypes}
\alias{TCGAquery_MultiSampleTypes}
\title{TCGAquery_MultiSampleTypes}
\usage{
TCGAquery_MultiSampleTypes(barcode, typesample)
}
\arguments{
\item{barcode}{barcode is a list of samples as TCGA barcodes}

\item{typesample}{is the type of samples barcode.
#' \tabular{lllll}{
TP   \tab TR \tab TB \tab TRBM \tab TAP \cr
TM \tab TAM \tab THOC \tab TBM \tab NB \cr
NT \tab NBC \tab NEBV \tab NBM \tab CELLC \cr
TRB \tab CELL \tab XP \tab XCL \tab
}}
}
\value{
a list of samples / barcode filtered by type sample selected
}
\description{
TCGAquery_MultiSampleTypes for a given list of samples and a type sample, return the samples that are
  from that type. TCGAquery_MultiSampleTypes allows using several type sample together.
}
\examples{
# selection of normal samples "NT"
dataFilt <- "TCGA-06-0125-01A-01D-A45W-05"
samplesNT <- TCGAquery_MultiSampleTypes(dataFilt, typesample = c("NT"))
}

