% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_Social.R
\name{TCGAquery_Social}
\alias{TCGAquery_Social}
\title{Finds the number of downloads of a package on CRAN or BIOC and
find questions in website ("bioconductor.org", "biostars.org", "stackoverflow).}
\usage{
TCGAquery_Social(siteToFind = NULL, listPackage = NULL, KeyInfo = NULL)
}
\arguments{
\item{siteToFind}{website ("bioconductor.org", "biostars.org", "stackoverflow)
related to TCGA or a package}

\item{listPackage}{list of package to find in bioconductor}

\item{KeyInfo}{is a key to find in biostars related to TGGA or package}
}
\value{
table with number of downloads about a package
}
\description{
Finds the number of downloads of a package on CRAN or BIOC
}
\examples{
TCGAquery_Social("bioconductor.org","BiocCheck")
}

