% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{TCGAquery_subtype}
\alias{TCGAquery_subtype}
\title{Retrieve molecular subtypes for a given tumor}
\usage{
TCGAquery_subtype(tumor)
}
\arguments{
\item{tumor}{is a cancer Examples:
\tabular{lllll}{
lgg   \tab gbm \tab luad \tab stad \tab brca\cr
coad \tab read \tab  \tab  \tab
}}
}
\value{
a data.frame with barcode and molecular subtypes
}
\description{
TCGAquery_subtype Retrieve molecular subtypes for a given tumor
}
\examples{
dataSubt <- TCGAquery_subtype(tumor = "lgg")
}

