% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_subtypes.R
\name{TCGAquery_subtypes}
\alias{TCGAquery_subtypes}
\title{Retrieve multiple molecular subtypes according last publications}
\usage{
TCGAquery_subtypes(tumor = NULL, path = ".")
}
\arguments{
\item{tumor}{is character such as cancer from TCGA. Examples:
\tabular{lllll}{
OV   \tab BRCA \tab CESC \tab ESCA \tab PCPG\cr
LUSC \tab LGG  \tab SKCM \tab KICH \tab CHOL\cr
GBM  \tab UCEC \tab PRAD \tab PAAD \tab THYM\cr
KIRC \tab THCA \tab SARC \tab LAML \tab TGCT\cr
COAD \tab KIRP \tab HNSC \tab ACC  \tab UVM \cr
READ \tab BLCA \tab DLBC \tab UCS  \tab FPPP\cr
LUAD \tab LIHC \tab STAD \tab MESO \tab CNTL
}}

\item{path}{Directory to save the downloaded data}
}
\value{
data.frame with information about molecular cancer subtypes
}
\description{
Retrieve multiple molecular subtypes according last publications
}
\examples{
library(xlsx)
GBM_path_subtypes <- TCGAquery_subtypes(tumor = "gbm",path ="dataGBM")
GBM_subtypes <- read.xlsx2(GBM_path_subtypes,1,stringsAsFactors = NULL)
}

