% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAsocial.R
\name{TCGAsocial}
\alias{TCGAsocial}
\title{TCGAsocial}
\usage{
TCGAsocial(listPackage)
}
\arguments{
\item{listPackage}{list of package to find}
}
\value{
table with number of downloads about a package
}
\description{
Find number of downloads about a package in CRAN or BioC
}

