% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAupdate.R, R/TCGAupdate2.R
\name{TCGAUpdate}
\alias{TCGAUpdate}
\title{TCGAUpdate}
\usage{
TCGAUpdate()

TCGAUpdate()
}
\description{
replace the dataFolders.rda matrix with the
            new matrix from "The Cancer Genome Atlas" ftp

replace the dataFolders.rda matrix with the
            new matrix from "The Cancer Genome Atlas" ftp
}
\author{
Davide

Davide
}
\seealso{
TCGAQuery

TCGAQuery
}

