% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGA_biolinks_functions.R
\name{TCGAVersion}
\alias{TCGAVersion}
\title{TCGA Download}
\usage{
TCGAVersion(Tumor, PlatformType, PlatformAndAssociatedData)
}
\arguments{
\item{Tumor}{a character string indicating the cancer type for
which to download data. Options include ACC, BLCA, BRCA,
CESC, COAD, DLBC, ESCA, GBM, HNSC, KICH, KIRC, KIRP, LAML,
LGG, LIHC, LUAD, LUSC, OV, PAAD, PRAD, READ, SARC, SKCM, STAD,
THCA, UCEC, UCS. Look at https://tcga-data.nci.nih.gov/tcga/
for Available Cancer Types.}

\item{PlatformType}{"illuminahiseq_rnaseq"}

\item{PlatformAndAssociatedData}{data frame 615 observations of 12 variables,
indicating the different characteristics of the data
e.g. tumour, type, species.}
}
\description{
TCGA Download
}

