% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{TCGAvisualize_BarPlot}
\alias{TCGAvisualize_BarPlot}
\title{Barplot of subtypes and clinical info in groups of gene expression clustered.}
\usage{
TCGAvisualize_BarPlot(DFfilt, DFclin, DFsubt, data_Hc2, Subtype, cbPalette,
  filename, width, height, dpi)
}
\arguments{
\item{DFfilt}{write}

\item{DFclin}{write}

\item{DFsubt}{write}

\item{data_Hc2}{write}

\item{Subtype}{write}

\item{cbPalette}{Define the colors of the bar.}

\item{filename}{The name of the pdf file}

\item{width}{Image width}

\item{height}{Image height}

\item{dpi}{Image dpi}
}
\value{
barplot image in pdf or png file
}
\description{
Barplot of subtypes and clinical info in groups of gene expression clustered.
}
