% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{TCGAvisualize_meanMethylation}
\alias{TCGAvisualize_meanMethylation}
\title{Mean methylation boxplot}
\usage{
TCGAvisualize_meanMethylation(data, groupCol = NULL, subgroupCol = NULL,
  shapes = NULL, print.pvalue = FALSE,
  filename = "sampleMeanMethylationByGroups.pdf",
  ylab = expression(paste("Mean DNA methylation (", beta, "-values)")),
  xlab = NULL, title = "Mean DNA methylation", labels = NULL,
  group.legend = NULL, subgroup.legend = NULL, color = c("green", "red",
  "purple", "orange", "salmon", "grey"))
}
\arguments{
\item{data}{SummarizedExperiment object obtained from TCGAPrepare}

\item{groupCol}{Columns in colData(data) that defines the groups. If no
columns defined a columns called "Patients" will be used}

\item{subgroupCol}{Columns in colData(data) that defines the subgroups.}

\item{shapes}{Shape vector of the subgroups. It must have the size of the levels
of the subgroups. Example: shapes = c(21,23) if for two levels}

\item{print.pvalue}{Print p-value for two groups}

\item{filename}{The name of the pdf that will be saved}

\item{ylab}{y axis text in the plot}

\item{xlab}{x axis text in the plot}

\item{title}{main title in the plot}

\item{labels}{Labels of the groups}

\item{group.legend}{Name of the group legend. DEFAULT: groupCol}

\item{subgroup.legend}{Name of the subgroup legend. DEFAULT: subgroupCol}

\item{color}{vector of colors to be used in graph}
}
\value{
Save the pdf survival plot
}
\description{
Creates a mean methylation boxplot divided in by groups
}
\examples{
nrows <- 200; ncols <- 21
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GenomicRanges::GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                   IRanges::IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                    strand=sample(c("+", "-"), 200, TRUE),
                    feature_id=sprintf("ID\%03d", 1:200))
colData <- S4Vectors::DataFrame(Treatment=rep(c("ChIP", "Input","Other"), 7),
                    row.names=LETTERS[1:21],
                    group=rep(c("group1","group2","group3"),c(7,7,7)))
data <- SummarizedExperiment::SummarizedExperiment(
         assays=S4Vectors::SimpleList(counts=counts),
         rowRanges=rowRanges,
         colData=colData)
TCGAvisualize_meanMethylation(data,groupCol  = "group")
}

