% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_mutation}
\alias{TCGAvisualize_mutation}
\title{Visualize mutation}
\usage{
TCGAvisualize_mutation(data = NULL, colors = NULL, samples = NULL,
  geneList = NULL, filename = NULL, threshold = 0, by = "genes",
  groupCol = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame with the cluters and subytpe of cancers}

\item{colors}{Vector of colors for the barplot}

\item{samples}{Samples to consider mutation}

\item{geneList}{List of genes to plot}

\item{filename}{Name of the file to save the plot, can be pdf, png, svg etc..}

\item{threshold}{plot only if number of mutated genes are higher than threshold}

\item{by}{type of plot. Options: "genes" and "cluster".
 By "genes" the axis will receive the genes
and how many samples are mutated. By "cluster" will show the mutations of
only one gene in the cluster.}

\item{groupCol}{Must be provided if by is set to "cluster"}

\item{na.rm}{Remove NA groups}
}
\value{
A plot
}
\description{
See % of genes mutated
}

