% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_oncoprint}
\alias{TCGAvisualize_oncoprint}
\title{Creating a oncoprint}
\usage{
TCGAvisualize_oncoprint(mut, genes, filename, color,
  annotation.position = "bottom", annotation, height,
  rm.empty.columns = FALSE, show.column.names = FALSE,
  show.row.barplot = TRUE, label.title = "Mutation", label.font.size = 16,
  rows.font.size = 16, dist.col = 0.5, dist.row = 0.5,
  row.order = FALSE, heatmap.legend.side = "bottom",
  annotation.legend.side = "bottom")
}
\arguments{
\item{mut}{A dataframe from the mutation annotation file (see TCGAquery_maf from TCGAbiolinks)}

\item{genes}{Gene list}

\item{filename}{name of the pdf}

\item{color}{named vector for the plot}

\item{annotation.position}{Position of the annotation "bottom" or "top"}

\item{annotation}{Matrix or data frame with the annotation.
Should have a column bcr_patient_barcode with the same ID of the mutation object}

\item{height}{pdf height}

\item{rm.empty.columns}{If there is no alteration in that sample, whether remove it on the oncoprint}

\item{show.column.names}{Show column names? Default: FALSE}

\item{show.row.barplot}{Show barplot annotation on rows?}

\item{label.title}{Title of the label}

\item{rows.font.size}{Size of the fonts}

\item{labels.font.size}{Size of the fonts}
}
\value{
A oncoprint plot
}
\description{
Creating a oncoprint
}
\examples{
mut <- TCGAbiolinks::TCGAquery_maf(tumor = "GBM",
       archive.name = "ucsc.edu_GBM.IlluminaGA_DNASeq_automated.Level_2.1.1.0")
TCGAvisualize_oncoprint(mut = mut, genes = mut$Hugo_Symbol[1:10], rm.empty.columns = TRUE)
TCGAvisualize_oncoprint(mut = mut, genes = mut$Hugo_Symbol[1:10],
                 filename = "onco.pdf",
                 color=c("background"="#CCCCCC","DEL"="purple","INS"="yellow","SNP"="brown"))
clin <- TCGAbiolinks::TCGAquery_clinic("gbm","clinical_patient")
clin <- clin[,c("bcr_patient_barcode","disease","gender","tumor_tissue_site","race","vital_status")]
TCGAvisualize_oncoprint(mut = mut, genes = mut$Hugo_Symbol[1:20],
                filename = "onco.pdf",
                annotation = clin,
                color=c("background"="#CCCCCC","DEL"="purple","INS"="yellow","SNP"="brown"),
                rows.font.size=10,
                heatmap.legend.side = "right",
                dist.col = 0,
                label.font.size = 10)

}

