% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_PCA}
\alias{TCGAvisualize_PCA}
\title{Principal components analysis (PCA) plot}
\usage{
TCGAvisualize_PCA(dataFilt, dataDEGsFiltLevel, ntopgenes)
}
\arguments{
\item{dataFilt}{A filtered dataframe or numeric matrix where each row represents a gene,
each column represents a sample from function TCGAanalyze_Filtering}

\item{dataDEGsFiltLevel}{table with DEGs, log Fold Change (FC), false discovery rate (FDR),
the gene expression level, etc, from function TCGAanalyze_LevelTab.}

\item{ntopgenes}{number of DEGs genes to plot in PCA}
}
\value{
principal components analysis (PCA) plot of PC1 and PC2
}
\description{
TCGAvisualize_PCA performs a principal components analysis (PCA) on the given data matrix
  and returns the results as an object of class prcomp, and shows results in PCA level.
}
\examples{
# normalization of genes
dataNorm <- TCGAbiolinks::TCGAanalyze_Normalization(dataBRCA, geneInfo)
# quantile filter of genes
dataFilt <- TCGAanalyze_Filtering(dataNorm, 0.25)
# Principal Component Analysis plot for ntop selected DEGs
TCGAvisualize_PCA(dataFilt,dataDEGsFiltLevel, ntopgenes = 200)
}

