% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_profilePlot}
\alias{TCGAvisualize_profilePlot}
\title{Profile plot}
\usage{
TCGAvisualize_profilePlot(data = NULL, groupCol = NULL, subtypeCol = NULL,
  colors = NULL, filename = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A data frame with the cluters and subytpe of cancers}

\item{groupCol}{Names of tre columns with the cluster information}

\item{subtypeCol}{Name of the column with the subtype information}

\item{filename}{Name of the column with the subtype information}

\item{na.rm}{Remove NA groups? Default = FALSE}
}
\value{
A plot
}
\description{
Displaty the association between cancer subtypes and any kind of clustering.
}
\examples{
query <- TCGAquery(tumor = "lgg")
\dontrun{
clin <- TCGAquery_clinic("lgg","clinical_patient")
TCGAvisualize_profilePlot ()
}
}

