% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{TimeUse}
\alias{TimeUse}
\title{TimeUse}
\usage{
TimeUse(func)
}
\arguments{
\item{func}{func}
}
\value{
time execution of a function
}
\description{
TimeUse
}
\examples{
TimeUse(ansEA <- EAcomplete(TFname="DEA genes Normal Vs Tumor",
 Genelist))
}

