% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{UseRaw_afterFilter}
\alias{UseRaw_afterFilter}
\title{Use raw count from the DataPrep object which genes are removed by normalization and filtering steps.}
\usage{
UseRaw_afterFilter(DataPrep, DataFilt)
}
\arguments{
\item{DataPrep}{DataPrep object returned by TCGAanalyze_Preprocessing()}

\item{DataFilt}{Filtered data frame containing samples in columns and genes in rows after normalization and/or filtering steps}
}
\value{
Filtered return object similar to DataPrep with genes removed after normalization and filtering process.
}
\description{
function to keep raw counts after filtering and/or normalizing.
}
\examples{
\dontrun{
  dataPrep_raw <- UseRaw_afterFilter(dataPrep, dataFilt)
}
}
