% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(data, groupCol = NULL, group1 = NULL, group2 = NULL,
  filename = "volcano.pdf", ylab = "- 1*log10 of the Significance",
  xlab = "DNA Methylation", title = "Volcano Plot TCGA GBM Tumors",
  legend = "Legend", color = c(`1` = "black", `2` = "green", `3` = "red"),
  label = c(`1` = "Not Significant", `2` = "Hypermethylated", `3` =
  "Hypomethylated"), xlim = NULL, ylim = NULL, p.cut = 0.05,
  diffmean.cut = 0)
}
\arguments{
\item{data}{SummarizedExperiment obtained from the TCGAPrepare}

\item{groupCol}{Columns with the groups inside the SummarizedExperiment
object. (This will be obtained by the function colData(data))}

\item{group1}{In case our object has more than 2 groups, you should set
the name of the group}

\item{group2}{In case our object has more than 2 groups, you should set
the name of the group}

\item{filename}{pdf filename}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title}

\item{legend}{legend title}

\item{color}{vector of colors to be used in graph}

\item{label}{vector of labels to be used in the figure}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}

\item{p.cut}{p values threshold}

\item{diffmean.cut}{diffmean threshold}
}
\value{
A dataframe with the Composite.Element.REF and
        the group it was classified
         group 1 = Not Significant
         group 2 = Hypermethylated
         group 3 = Hypomethylated
}
\description{
In order to searches for the probes that are different methylated and
  are significant, we use a volcano plot (x-axis:diff mean methylation,
  y-axis: significance) that compares the methylation data between the two groups.
  Firstly, it calculates the difference between the mean methylation of each group
for each probes. After, it calculates the p-value using the wilcoxon test
using the Benjamini-Hochberg adjustment method. With both values, it is possible
to analyse the data.
}
\examples{
nrows <- 200; ncols <- 20
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GenomicRanges::GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                   IRanges::IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                    strand=sample(c("+", "-"), 200, TRUE),
                    feature_id=sprintf("ID\%03d", 1:200))
colData <- S4Vectors::DataFrame(Treatment=rep(c("ChIP", "Input"), 5),
                    row.names=LETTERS[1:20],
                    group=rep(c("group1","group2"),c(10,10)))
data <- SummarizedExperiment::SummarizedExperiment(
         assays=S4Vectors::SimpleList(counts=counts),
         rowRanges=rowRanges,
         colData=colData)
hypo.hyper <- volcanoPlot(data, p.cut = 0.85)
}

