% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{calculate.pvalues}
\alias{calculate.pvalues}
\title{Calculate pvalues}
\usage{
calculate.pvalues(values, idx1 = NULL, idx2 = NULL, paired = TRUE,
  exact = TRUE)
}
\arguments{
\item{values}{Dataframe with values}

\item{idx1}{Index of the values in group 1}

\item{idx2}{Index of the values in group 2}

\item{paired}{Do a paired wilcoxon test? Default: True}

\item{exact}{Do a exact wilcoxon test? Default: True}
}
\value{
Data frame with cols p values/p values adjusted

Data frame with two cols
        p-values/p-values adjusted
}
\description{
Calculate pvalues
}
\details{
Input: two matrix to be compared with wilcoxon test,
          a boolean value to do a paired or non-paired test
   Output: p-values (non-adj/adj) histograms, p-values (non-adj/adj)
}
\examples{
patient <- paste("patient",1:100)
probes <- paste0("cg00",1:100)
beta.values <- runif(10000,0,1)
beta.values <- matrix(beta.values,ncol = 100,nrow = 100)
values <- as.data.frame(beta.values)
rownames(values) <- patient
colnames(values) <- probes
pvalues <- calculate.pvalues(values,1:50,51:100)
}

