% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{chol_maf}
\alias{chol_maf}
\title{TCGA CHOL MAF transformed to maftools obejct}
\format{An object of class  MAF}
\description{
TCGA CHOL MAF transformed to maftools obejct
}
\keyword{internal}
