% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{clinic}
\alias{clinic}
\title{clinic}
\usage{
clinic(cancer, clinical_data_type)
}
\arguments{
\item{cancer}{a character vector indicating cancer type Examples:
\tabular{lllll}{
OV   \tab BRCA \tab CESC \tab ESCA \tab PCPG\cr
LUSC \tab LGG  \tab SKCM \tab KICH \tab CHOL\cr
GBM  \tab UCEC \tab PRAD \tab PAAD \tab THYM\cr
KIRC \tab THCA \tab SARC \tab LAML \tab TGCT\cr
COAD \tab KIRP \tab HNSC \tab ACC  \tab UVM \cr
READ \tab BLCA \tab DLBC \tab UCS  \tab FPPP\cr
LUAD \tab LIHC \tab STAD \tab MESO \tab CNTL
}
For information about cancer types: https://tcga-data.nci.nih.gov/tcga/}

\item{clinical_data_type}{a character vector indicating the types of
clinical data Example:
\tabular{ll}{
biospecimen_aliquot \tab biospecimen_analyte \cr
biospecimen_cqcf \tab biospecimen_diagnostic_slides \cr
biospecimen_normal_control \tab biospecimen_portion \cr
biospecimen_protocol \tab biospecimen_sample \cr
biospecimen_shipment_portion \tab biospecimen_slide \cr
biospecimen_tumor_sample \tab clinical_cqcf \cr
clinical_drug \tab clinical_follow_up_v1.5 \cr
clinical_follow_up_v2.1 \tab clinical_follow_up_v4.0 \cr
clinical_follow_up_v4.0_nte \tab clinical_nte \cr
clinical_omf_v4.0 \tab clinical_patient \cr
clinical_radiation
}}
}
\value{
clinic data
}
\description{
clinic
}
\examples{
data <- clinic("LGG","clinical_drug")
}

