% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{colDataPrepare}
\alias{colDataPrepare}
\title{Create samples information matrix for GDC samples}
\usage{
colDataPrepare(barcode)
}
\arguments{
\item{barcode}{TCGA or TARGET barcode}
}
\description{
Create samples information matrix for GDC samples add subtype information
}
\examples{
\dontrun{
  query.met <- GDCquery(project = c("TCGA-GBM","TCGA-LGG"),
                        legacy = TRUE,
                        data.category = "DNA methylation",
                        platform = c("Illumina Human Methylation 450",
                                      "Illumina Human Methylation 27"))
  colDataPrepare(getResults(query.met)$cases)
}
}
