% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{diffmean}
\alias{diffmean}
\title{Calculate diffmean methylation between two groups}
\usage{
diffmean(data, groupCol = NULL, group2 = NULL, group1 = NULL)
}
\arguments{
\item{group2}{Data frame probes vs patient}

\item{group1}{Data frame probes vs patient}
}
\value{
dataframe with diffmean values
}
\description{
Calculate diffmean methylation between two groups
}
\examples{
nrows <- 200; ncols <- 20
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                   IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                    strand=sample(c("+", "-"), 200, TRUE),
                    feature_id=sprintf("ID\%03d", 1:200))
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 10),
                    row.names=LETTERS[1:20],
                    group=rep(c("group1","group2"),c(10,10)))
data <- SummarizedExperiment(assays=SimpleList(counts=counts),
                          rowRanges=rowRanges, colData=colData)
data <- diffmean(data)
}

