% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{diffmean}
\alias{diffmean}
\title{calculate diffmean methylation between two groups}
\usage{
diffmean(group1, group2)
}
\arguments{
\item{group1}{Data frame probes vs patient}

\item{group2}{Data frame probes vs patient}
}
\value{
dataframe with diffmean values
}
\description{
calculate diffmean methylation between two groups
}
\examples{
patient <- paste("patient",1:100)
probes <- paste0("cg00",1:100)
beta.values <- runif(10000,0,1)
beta.values <- matrix(beta.values,ncol = 100,nrow = 100)
values <- as.data.frame(beta.values)
colnames(values) <- patient
rownames(values) <- probes
diff.group1xgroup2 <- diffmean(values[,1:50],values[,51:100])
}

