% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{EAbarplot}
\alias{EAbarplot}
\title{EAbarplot}
\usage{
EAbarplot(tf, GOMFTab, GOBPTab, GOCCTab, PathTab, nBar, nRGTab)
}
\arguments{
\item{tf}{is a list of gene symbols}

\item{GOMFTab}{is results from EAcomplete related to Molecular Function (MF)}

\item{GOBPTab}{is results from EAcomplete related to Biological Process (BP)}

\item{GOCCTab}{is results from EAcomplete related to Cellular Component (CC)}

\item{PathTab}{is results from EAcomplete related to Pathways EA}

\item{nBar}{is the number of bar histogram selected to show (default = 10)}

\item{nRGTab}{is the gene signature list with gene symbols.}
}
\value{
Complete barPlot from Enrichment Analysis showing significant (default FDR < 0.01)
BP,CC,MF and pathways enriched by list of genes.
}
\description{
EAbarplot plots the result from EAcomplete in a complete barPlot
}
\examples{
Genelist <- c("FN1","COL1A1")
ansEA <- EAcomplete(TFname="DEA genes Normal Vs Tumor",Genelist)
EAbarplot(tf = rownames(ansEA$ResBP),
         GOBPTab = ansEA$ResBP,
         GOCCTab = ansEA$ResCC,
         GOMFTab = ansEA$ResMF,
        PathTab = ansEA$ResPat,
         nRGTab = Genelist,
         nBar = 10)
\dontrun{
Genelist <- rownames(dataDEGsFiltLevel)
system.time(ansEA <- EAcomplete(TFname="DEA genes Normal Vs Tumor",Genelist))
# Enrichment Analysis EA (TCGAVisualize)
# Gene Ontology (GO) and Pathway enrichment barPlot
EAbarplot(tf = rownames(ansEA$ResBP),
         GOBPTab = ansEA$ResBP,
         GOCCTab = ansEA$ResCC,
         GOMFTab = ansEA$ResMF,
        PathTab = ansEA$ResPat,
         nRGTab = Genelist,
         nBar = 10)
}
}

