% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{EAcomplete}
\alias{EAcomplete}
\title{EAcomplete}
\usage{
EAcomplete(TFname, RegulonList)
}
\arguments{
\item{TFname}{TFname}

\item{RegulonList}{RegulonList}
}
\value{
EAcomplete plot
}
\description{
EAcomplete
}
\examples{
\dontrun{

Genelist <- rownames(dataDEGsFiltLevel)
TimeUse(ansEA <- EAcomplete(TFname="DEA genes Normal Vs Tumor",Genelist))
}
}

