% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{EnrichmentAnalysis}
\alias{EnrichmentAnalysis}
\title{EnrichmentAnalysis}
\usage{
EnrichmentAnalysis(GeneName, RegulonList, TableEnrichment, EAGenes, GOtype,
  FDRThresh = 0.01)
}
\arguments{
\item{GeneName}{GeneName}

\item{RegulonList}{RegulonList}

\item{TableEnrichment}{TableEnrichment}

\item{EAGenes}{EAGenes}

\item{GOtype}{GOtype}

\item{FDRThresh}{FDRThresh}
}
\value{
EAcomplete plot
}
\description{
EnrichmentAnalysis
}
\examples{
\dontrun{

EAGenes <- get("EAGenes")
RegulonList <- rownames(dataDEGsFiltLevel)
DAVID_BP_matrix <- get("DAVID_BP_matrix")
ResBP <- EnrichmentAnalysis(GeneName="DEA genes Normal Vs Tumor",RegulonList,DAVID_BP_matrix,
                           EAGenes,GOtype = "DavidBP")
                           }
}

