% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{ER_status_BRCA}
\alias{ER_status_BRCA}
\title{ER_status_BRCA}
\usage{
ER_status_BRCA(barcode, ER, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{ER}{ER}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
ER_status_BRCA
}
\description{
ER_status_BRCA
}
\examples{
# clin <- clinic("BRCA","clinical_patient")
clin <- clinBRCA
ER_status_BRCA(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),
"Positive",clin)
}

