% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{GDCprepare}
\alias{GDCprepare}
\title{Prepare GDC data}
\usage{
GDCprepare(query, save = FALSE, save.filename, directory = "GDCdata",
  summarizedExperiment = TRUE, remove.files.prepared = FALSE)
}
\arguments{
\item{query}{A query for GDCquery function}

\item{save}{Save result as RData object?}

\item{save.filename}{Name of the file to be save if empty an automatic will be created}

\item{directory}{Directory/Folder where the data was downloaded. Default: GDCdata}

\item{summarizedExperiment}{Create a summarizedExperiment? Default TRUE (if possible)}

\item{remove.files.prepared}{Remove the files read? Default: FALSE
This argument will be considered only if save argument is set to true}
}
\value{
A summarizedExperiment or a data.frame
}
\description{
Reads the data downloaded and prepare it into an R object
}
\examples{
query <- GDCquery(project = "TCGA-KIRP",
                  data.category = "Simple Nucleotide Variation",
                  data.type = "Masked Somatic Mutation")
                  GDCdownload(query, method = "api", directory = "maf")
GDCdownload(query, method = "api", directory = "maf")
maf <- GDCprepare(query, directory = "maf")

query <- GDCquery(project = "TCGA-ACC",
                   data.category =  "Copy number variation",
                   legacy = TRUE,
                   file.type = "hg19.seg",
                   barcode = c("TCGA-OR-A5LR-01A-11D-A29H-01", "TCGA-OR-A5LJ-10A-01D-A29K-01"))
# data will be saved in  GDCdata/TCGA-ACC/legacy/Copy_number_variation/Copy_number_segmentation
GDCdownload(query, method = "api")
acc.cnv <- GDCprepare(query)
}

