% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{GDCPrepare}
\alias{GDCPrepare}
\title{Prepare GDC data}
\usage{
GDCPrepare(query, save = FALSE, save.filename, summarizedExperiment = TRUE)
}
\arguments{
\item{query}{A query for GDCquery function}

\item{save}{Save result as RData object?}

\item{save.filename}{Name of the file to be save if empty an automatic will be created}
}
\value{
A summarizedExperiment or a data.frame
}
\description{
Reads the data downloaded and prepare it into an R object
}

