% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{GDCquery_clinic}
\alias{GDCquery_clinic}
\title{Get GDC clinical data}
\usage{
GDCquery_clinic(project, type = "clinical", save.csv = FALSE)
}
\arguments{
\item{project}{A valid project (see list with getGDCprojects()$project_id)]}

\item{type}{A valid type. Options "clinical", "Biospecimen"  (see list with getGDCprojects()$project_id)]}

\item{save.csv}{Write clinical information into a csv document}
}
\value{
A data frame with the clinical information
}
\description{
GDCquery_clinic will download all clinical information from the API
as the one with using the button from each project
}
\examples{
clin <- GDCquery_clinic("TCGA-ACC", type = "clinical", save.csv = TRUE)
clin <- GDCquery_clinic("TCGA-ACC", type = "biospecimen", save.csv = TRUE)
}

