% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAQuery.R
\name{GDCquery_Maf}
\alias{GDCquery_Maf}
\title{Retrieve open access maf files from GDC server}
\usage{
GDCquery_Maf(tumor, save.csv = FALSE)
}
\arguments{
\item{tumor}{a valid tumor}

\item{save.csv}{Write maf file into a csv document}
}
\value{
A data frame with the maf file information
}
\description{
GDCquery_Maf uses the following guide to download maf files
  https://gdc-docs.nci.nih.gov/Data/Release_Notes/Data_Release_Notes/
}
\examples{
acc.maf <- GDCquery_Maf("ACC")
}

