% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{gender_BRCA}
\alias{gender_BRCA}
\title{gender_BRCA}
\usage{
gender_BRCA(barcode, gender, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{gender}{gender}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA
}
\description{
gender_BRCA
}
\examples{
\dontrun{
clin <- clinic("BRCA","clinical_patient")
gender_BRCA (c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),"FEMALE",clin)
}
}

