% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{geneInfo}
\alias{geneInfo}
\title{geneInfo for normalization}
\format{A data frame with 20531 rows and 2 variables}
\description{
geneInfo for normalization
}
\keyword{internal}

