% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{data}
\name{geneInfoHT}
\alias{geneInfoHT}
\title{geneInfoHT for normalization of HTseq data}
\format{A data frame with 23486 rows and 2 variables}
\description{
geneInfoHT for normalization of HTseq data
}
\keyword{internal}

