% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{GenesCutID}
\alias{GenesCutID}
\title{GenesCutID}
\usage{
GenesCutID(GeneList)
}
\arguments{
\item{GeneList}{GeneList}
}
\value{
list of gene symbol without IDs
}
\description{
GenesCutID
}
\examples{
GenesCutID(c("?|100133144","?|100134869","?|10357" ))
}

