% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAInternal.R
\name{GeneSplitRegulon}
\alias{GeneSplitRegulon}
\title{GeneSplitRegulon}
\usage{
GeneSplitRegulon(Genelist, Sep)
}
\arguments{
\item{Genelist}{Genelist}

\item{Sep}{Sep}
}
\value{
GeneSplitRegulon
}
\description{
GeneSplitRegulon
}

