% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get.GRCh.bioMart}
\alias{get.GRCh.bioMart}
\title{Get hg19 gene annotation or hg38 (gencode v36)}
\usage{
get.GRCh.bioMart(genome = c("hg19", "hg38"), as.granges = FALSE)
}
\arguments{
\item{genome}{hg38 or hg19}

\item{as.granges}{Output as GRanges or data.frame}
}
\description{
Get hg19 (from biomart) or hg38 (gencode v36) gene annotation
}
