% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getNbFiles}
\alias{getNbFiles}
\title{Get Number of files in GDC for a project}
\usage{
getNbFiles(project, data.category, legacy = FALSE)
}
\arguments{
\item{project}{A  GDC project}

\item{data.category}{A  GDC project data category}

\item{legacy}{Select between Harmonized or Legacy database}
}
\description{
Get Number of files in GDC for a project
}
\examples{
getNbFiles("TCGA-ACC","Clinical")
getNbFiles("CPTAC-2","Clinical")
}
