% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getProjectSummary}
\alias{getProjectSummary}
\title{Get Project Summary from GDC}
\usage{
getProjectSummary(project, legacy = FALSE)
}
\description{
Get Project Summary from GDC
}
\examples{
getProjectSummary("TCGA-ACC")
getProjectSummary("CPTAC-2")
}
