% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getProjectSummary}
\alias{getProjectSummary}
\title{Get Project Summary from GDC}
\usage{
getProjectSummary(project)
}
\arguments{
\item{project}{A  GDC project}
}
\description{
Get Project Summary from GDC
}
\examples{
getProjectSummary("TCGA-ACC")
\dontrun{
getProjectSummary("CPTAC-2")
}
}
\author{
Tiago Chedraoui Silva
}
