% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getSampleFilesSummary}
\alias{getSampleFilesSummary}
\title{Retrieve summary of files per sample in a project}
\usage{
getSampleFilesSummary(project)
}
\arguments{
\item{project}{A GDC project}
}
\value{
A data frame with the maf file information
}
\description{
Retrieve the numner of files under each
  data_category + data_type + experimental_strategy + platform
  Almost like https://portal.gdc.cancer.gov/exploration
}
\examples{
   summary <- getSampleFilesSummary("TCGA-LUAD")
\dontrun{
   summary <- getSampleFilesSummary(c("TCGA-OV","TCGA_ACC"))
}
}
