% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{HER_status_BRCA}
\alias{HER_status_BRCA}
\title{HER_status_BRCA}
\usage{
HER_status_BRCA(barcode, HER, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{HER}{HER}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
HER_status_BRCA
}
\description{
HER_status_BRCA
}
\keyword{internal}

