% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{MatchedCoupledSampleTypes}
\alias{MatchedCoupledSampleTypes}
\title{MatchedCoupledSampleTypes}
\usage{
MatchedCoupledSampleTypes(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
a list of samples / barcode filtered by type sample selected
}
\description{
MatchedCoupledSampleTypes
}
\examples{
MatchedCoupledSampleTypes(c("TCGA-B0-4698-01Z-00-DX1",
                             "TCGA-B0-4698-02Z-00-DX1"),
                             c("TP","TR"))
}

