% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{metMeanBoxplot}
\alias{metMeanBoxplot}
\title{Mean methylation boxplot}
\usage{
metMeanBoxplot(data, sort = FALSE, filename = "G-CIMP-mean.methylation.pdf",
  ylab = "Mean DNA methylation", xlab = "DNA Methylation Clusters",
  title = "Mean DNA methylation by cluster", legend = "Legend",
  color = c("green", "red", "purple", "orange", "salmon", "grey"))
}
\arguments{
\item{data}{data frame first col mean methylation by patient, second groups}

\item{sort}{Sort by mean methylation? False by default}

\item{filename}{pdf filename}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title}

\item{legend}{legend title}

\item{color}{vector of colors to be used in graph}
}
\value{
Save the survival plot
}
\description{
Mean methylation boxplot.
  Input: a dataframe with two columns:
   1 - values of mean methylation
   2 - Groups it belongs
}
\examples{
# -----------------
# |      data       |
# -----------------
# | mean.met|cluster|
# -------------------
  avg <- runif(500,0,1)
  cluster <- c('Lgm1','Lgm2','Lgm3','Lgm4','Lgm5','Lgm6')
  cluster <- sample(cluster, 500,replace = TRUE)
  data <- data.frame(avg,cluster)
  metMeanBoxplot(data)
}

