% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{organizeMethylationMetaDataFrame}
\alias{organizeMethylationMetaDataFrame}
\title{organize TCGA Methylation MetaData}
\usage{
organizeMethylationMetaDataFrame(wd = NULL)
}
\arguments{
\item{wd}{Directory with the files}
}
\value{
\code{invisible (metadata)}
}
\description{
Organize TCGA methylation metadata for the mean methylation analysis.
}
\examples{
#-------------------------- Preparing data and metadata
# met: data frame with probe info (probe id, ge) and beta values
# beta: data frame with only beta values (probe vs patient)
# probe: data frame with only probe info(probe vs patient)
# Conclusion: met = probe +d beta
# met.md = patient metadata
query <- tcgaQuery(tumor = "GBM", platform = "HumanMethylation450", level = 3)
tcgaDownload(query[1,],path = "data")
query <- tcgaQuery(tumor = "GBM", platform = "bio", level = 2)
tcgaDownload(query,path = "data")

met <- organizeMethylationDataFrame("data")
probe <- met[,1:4]
beta  <- met[,5:ncol(met)]

met.md <- organizeMethylationMetaDataFrame("data")
samples <- colnames(beta)
idx <- is.element(met.md$bcr_patient_barcode,strtrim(samples,12))
met.md <- met.md[idx,]

# random split of pacients into groups
met.md$cluster <- c(rep("group1",nrow(met.md) / 4),
                    rep("group2",nrow(met.md) / 4),
                    rep("group3",nrow(met.md) / 4),
                    rep("group4",nrow(met.md) - 3 * (floor(nrow(met.md) / 4))))
rownames(met.md) <- met.md$bcr_patient_barcode

#---------------------- survival
survivalPlot(met.md)

#----------------------mean methylation
met.mean <- data.frame(apply(beta, 2, mean, na.rm = TRUE))
colnames(met.mean) <- "avg"
met.mean$patient <-  strtrim(rownames(met.mean),12)
aux <- merge(met.mean,met.md,
             by.x = "patient",
             by.y = "bcr_patient_barcode")
metMeanBoxplot(aux)

#----------------------- calculate.pvalues (just an example)
beta.t <-  data.frame(t(beta))
# TODO verify class of the object
pvalues <- calculate.pvalues(beta.t,c(1,3,5,7,9,11),c(2,4,6,8,10,12))

#------------- volcano plot
probe$p.value <- pvalues[,1]
probe$p.value.adj <- pvalues[,2]
prim.rec <- diffmean(beta,c(1,3,5,7,9,11),c(2,4,6,8,10,12))
probe <- cbind(probe,prim.rec)
hypo.hyper <- volcanoPlot(probe,p.cut = 0.619)
}

