% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{plotPCAforGroups}
\alias{plotPCAforGroups}
\title{plotPCAforGroups}
\usage{
plotPCAforGroups(dataFilt, dataDEGsFiltLevel, ntopgenes)
}
\arguments{
\item{dataFilt}{dataFilt}

\item{dataDEGsFiltLevel}{dataDEGsFiltLevel}

\item{ntopgenes}{number of DEGs genes to plot in PCA}
}
\value{
PCA plot
}
\description{
plotPCAforGroups
}
\examples{
# normalization of genes
dataNorm <- TCGAbiolinks::RnaSeqNormalization(dataBRCA, geneInfo)
# quantile filter of genes
dataFilt <- RnaSeqFilt(dataNorm, 0.25)
# Principal Component Analysis plot for ntop selected DEGs
plotPCAforGroups(dataFilt,dataDEGsFiltLevel, ntopgenes = 200)
}

