% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{PR_status_BRCA}
\alias{PR_status_BRCA}
\title{PR_status_BRCA}
\usage{
PR_status_BRCA(barcode, PR, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{PR}{PR}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
PR_status_BRCA
}
\description{
PR_status_BRCA
}
\examples{
clin <- clinic("BRCA","clinical_patient")
PR_status_BRCA(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),
"Positive",clin)
}

