% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{RnaSeqFilt}
\alias{RnaSeqFilt}
\title{RnaSeqFilt}
\usage{
RnaSeqFilt(TableRnaseq, QuantileThresh)
}
\arguments{
\item{TableRnaseq}{is a dataframe or numeric matrix, each row represents a gene,
each column represents a sample come from TCGAPrepare}

\item{QuantileThresh}{is threshold selected as mean for filtering}
}
\value{
A filtered dataframe or numeric matrix where each row represents a gene,
each column represents a sample
}
\description{
RnaSeqFilt allows user to filter mRNA transcripts and miRNA,
   selecting a threshold. For istance returns all mRNA or miRNA with mean across all
   samples, higher than the threshold defined quantile mean across all samples.
}
\examples{
dataNorm <- TCGAbiolinks::RnaSeqNormalization(dataBRCA, geneInfo)
dataFilt <- RnaSeqFilt(dataNorm, 0.25)
}

