% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{RnaSeqNormalization}
\alias{RnaSeqNormalization}
\title{RnaSeqNormalization}
\usage{
RnaSeqNormalization(TCGA_RnaseqTable, geneInfo)
}
\arguments{
\item{TCGA_RnaseqTable}{TCGA_RnaseqTable}

\item{geneInfo}{geneInfo}
}
\value{
table normalized
}
\description{
RnaSeqNormalization
}

