% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{RnaSeqNormalization}
\alias{RnaSeqNormalization}
\title{RnaSeqNormalization}
\usage{
RnaSeqNormalization(TCGA_RnaseqTable, geneInfo)
}
\arguments{
\item{TCGA_RnaseqTable}{TCGA_RnaseqTable}

\item{geneInfo}{geneInfo}
}
\value{
table normalized
}
\description{
RnaSeqNormalization allows user to normalize mRNA transcripts and miRNA,
   using EDASeq package. Normalization for RNA-Seq Numerical and graphical summaries of RNA-Seq read data. Within-lane normalization procedures
   to adjust for GC-content effect (or other gene-level effects) on read counts:
   loess robust local regression, global-scaling, and full-quantile normalization
   (Risso et al., 2011). Between-lane normalization procedures to adjust for
   distributional differences between lanes (e.g., sequencing depth): global-scaling and full-quantile normalization (Bullard et al., 2010).
   For istance returns all mRNA or miRNA with mean across all
   samples, higher than the threshold defined quantile mean across all samples.
}
\examples{
dataNorm <- TCGAbiolinks::RnaSeqNormalization(dataBRCA, geneInfo)
}

