% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{SampleTypes}
\alias{SampleTypes}
\title{SampleTypes}
\usage{
SampleTypes(barcode, typesample)
}
\arguments{
\item{barcode}{barcode}

\item{typesample}{typesample}
}
\value{
SampleTypes
}
\description{
SampleTypes
}

