% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical.R
\name{stage_BRCA}
\alias{stage_BRCA}
\title{stage_BRCA}
\usage{
stage_BRCA(barcode, stage, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{stage}{stage}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA
}
\description{
stage_BRCA
}
\examples{
clin <- clinic("BRCA","clinical_patient")
stage_BRCA(c("TCGA-3C-AALK","TCGA-A2-A04Q","TCGA-A4-A04Q"),"stage_IX",clin)
}

