% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_clinical.R
\name{stage_BRCA}
\alias{stage_BRCA}
\title{stage_BRCA}
\usage{
stage_BRCA(barcode, stage, clinical_patient_data)
}
\arguments{
\item{barcode}{barcode}

\item{stage}{stage}

\item{clinical_patient_data}{clinical_patient_data}
}
\value{
stage_BRCA
}
\description{
stage_BRCA
}
\keyword{internal}

