% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{starbursAnalysis}
\alias{starbursAnalysis}
\title{Preparing methylation data and expression data}
\usage{
starbursAnalysis(met, expression)
}
\arguments{
\item{met}{methylation data}

\item{expression}{expression data}
}
\value{
Dataframe with methylation and expression data
}
\description{
Preparing methylation data and expression data
}
\examples{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAQuery(tumor = "GBM",samples = sample, level = 3)
TCGADownload(query,path = "exampleData",samples = sample, quiet = TRUE)
met <- TCGAPrepare(query, dir="exampleData")
met$diffmean <- runif(nrow(met),-3,3)
FDR <- runif(nrow(met),0,1)
p.value.adj <- runif(nrow(met),0,1)
GeneSymbol <- met$Gene_Symbol
dm <-  runif(nrow(met),-3,3)
expression <- data.frame(GeneSymbol,FDR,p.value.adj,dm)
gene.met <- starbursAnalysis(met,expression)
}

