% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{starburstPlot}
\alias{starburstPlot}
\title{Create starburst plot}
\usage{
starburstPlot(data, filename = "volcano.pdf",
  ylab = paste0("Gene Expression\\nlog10 of the",
  "adjusted Significance (FDR)"),
  xlab = paste0("DNA Methylation\\nlog10 of the",
  " adjusted Significance (FDR)"), title = "Starburst Plot",
  legend = "Methylation/Expression Relation", color = c(`1` = "black", `2` =
  "purple", `3` = "darkgreen", `4` = "blue", `5` = "darkred", `6` = "red", `7` =
  "green", `8` = "yellow", `9` = "orange"), label = c(`1` = "Not Significant",
  `2` = "Up & Hypo", `3` = "Down & Hypo", `4` = "hypo", `5` = "hyper", `6` =
  "Up", `7` = "Down", `8` = "Up & Hyper", `9` = "Down & Hyper"), xlim = NULL,
  ylim = NULL, p.cut = 0.05)
}
\arguments{
\item{data}{data frame columns: diffmean, p.value.adj}

\item{filename}{pdf filename}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title}

\item{legend}{legend title}

\item{color}{vector of colors to be used in graph}

\item{label}{vector of labels to be used in graph}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}

\item{p.cut}{p value cut-off}
}
\value{
Save a starburst plot
}
\description{
Create starburst plot
}
\details{
Input: data with gene expression/methylation expression
   Output: starburst plot
}
\examples{
\dontrun{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAQuery(tumor = "GBM",samples = sample, level = 3)
TCGADownload(query,path = "exampleData",samples = sample, quiet = TRUE)
met <- TCGAPrepare(query, dir="exampleData")
met <-  met[ met$Chromosome != "X" &  met$Chromosome != "Y", ]
met <- na.omit(met)
met$diffmean <- runif(nrow(met),-3,3)
FDR <- runif(nrow(met),0,1)
p.value.adj <- runif(nrow(met),0,1)
GeneSymbol <- met$Gene_Symbol
dm <-  runif(nrow(met),-3,3)
expression <- data.frame(GeneSymbol,FDR,p.value.adj,dm)
gene.met <- starbursAnalysis(met,expression)
starburstPlot(gene.met)
}
}

