% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SurvivalKMunivariate.R
\name{SurvivalKMunivariate}
\alias{SurvivalKMunivariate}
\title{SurvivalKMtable}
\usage{
SurvivalKMunivariate(clinical_patient, dataGE, Genelist, Survresult,
  ThreshTop = 0.67, ThreshDown = 0.33)
}
\arguments{
\item{clinical_patient}{clinical_patient}

\item{dataGE}{dataGE}

\item{Genelist}{Genelist}

\item{Survresult}{Survresult}

\item{ThreshTop}{ThreshTop}

\item{ThreshTop}{ThreshTop}
}
\value{
table with survival genes pvalues from KM of pubmed related to tfs.
}
\description{
Survival table with univariate Kaplan-Meier
}

