% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{survivalPlot}
\alias{survivalPlot}
\title{creates survival analysis}
\usage{
survivalPlot(met.md, legend = "Legend", cutoff = 0,
  main = "Kaplan-Meier Overall Survival Curves",
  ylab = "PROBABILITY OF SURVIVAL", xlab = "TIME SINCE DIAGNOSIS",
  time.reference = "DAYS", filename = "survival.pdf", color = c("green",
  "firebrick4", "orange3", "blue"), default.plot = "ggplot")
}
\arguments{
\item{met.md}{Data frame with the following columns: vital, os, cluster}

\item{legend}{Legend title of the figure}

\item{cutoff}{xlim}

\item{main}{main title}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{time.reference}{In order not to forget -> MONTHS DAY YEAR}

\item{filename}{output file name}

\item{color}{not implemented yet}

\item{default.plot}{You can use ggplot or plot}
}
\value{
Survival plot
}
\description{
Creates survival analysis
}

