% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAanalyze_added.R
\name{TCGAanalyze_DEA_Affy}
\alias{TCGAanalyze_DEA_Affy}
\title{Differentially expression analysis (DEA) using limma package.}
\usage{
TCGAanalyze_DEA_Affy(Pdatatable, PathFolder, FC.cut = 0.01)
}
\arguments{
\item{Pdatatable}{}

\item{PathFolder}{}

\item{FC.cut}{}
}
\value{
List of list with tables in 2 by 2 comparison
of the top-ranked genes from a linear model fitted by DEA's limma
}
\description{
Differentially expression analysis (DEA) using limma package.
}
\examples{
\dontrun{
to add example
}
}

