% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAbiolinks.R
\docType{package}
\name{TCGAbiolinks}
\alias{TCGAbiolinks}
\alias{TCGAbiolinks-package}
\title{Download data of samples from TCGA}
\description{
TCGAbiolinks allows you to Download data of samples from TCGA
}
\details{
The functions you're likely to need from \pkg{TCGAbiolinks} is
\code{\link{TCGADownload}}, \code{\link{TCGAQuery}} and
\code{\link{TCGAVersion}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

