% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGADownload}
\alias{TCGADownload}
\title{TCGA Download}
\usage{
TCGADownload(data = NULL, path = ".")
}
\arguments{
\item{data}{TCGASearch output}

\item{path}{location of the final data saving}
}
\value{
Download tcga into path
}
\description{
Download data previously selected using the TCGASeach
}
\examples{
\dontrun{
   tcgaDownload(data,'folder')
}
}
\seealso{
TCGASearch
}

