% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGADownload}
\alias{TCGADownload}
\title{TCGA Download}
\usage{
TCGADownload(data = NULL, path = ".", type = NULL, samples = NULL,
  quiet = FALSE, force = FALSE)
}
\arguments{
\item{data}{TCGAQuery output}

\item{path}{Directory to save the downloaded data}

\item{type}{Filter the files that will be downloaded by type}

\item{samples}{List of samples to download}

\item{quiet}{Supress output messages?. Default: FALSE}

\item{force}{Download files even if it was already downladed? Default: FALSE}
}
\value{
Download tcga into path
}
\description{
Download data previously selected using the TCGASeach
}
\examples{
samples <- c("TCGA-26-1442-01A-01R-1850-01")
   query <- TCGAQuery(tumor = "gbm", platform = "IlluminaHiSeq_RNASeqV2",
   level = "3", samples = samples)
   TCGADownload(query,path = "dataDemo2",samples = samples,
                type ="rsem.genes.results")
}
\seealso{
TCGAQuery
}

