% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGADownload}
\alias{TCGADownload}
\title{TCGA Download}
\usage{
TCGADownload(data = NULL, path = ".", type = NULL)
}
\arguments{
\item{data}{TCGASearch output}

\item{path}{location of the final data saving}

\item{type}{Get files with type pattern instead of downloading all the folder}
}
\value{
Download tcga into path
}
\description{
Download data previously selected using the TCGASeach
}
\examples{
\dontrun{
   tcgaDownload(data,path='folder',type="rsem.isoforms.results")
}
}
\seealso{
TCGASearch
}

