% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGADownload}
\alias{TCGADownload}
\title{TCGA Download}
\usage{
TCGADownload(data = NULL, path = ".", type = NULL, samples = NULL)
}
\arguments{
\item{data}{TCGASearch output}

\item{path}{location of the final data saving}

\item{type}{Get files with type pattern instead of downloading all the folder}
}
\value{
Download tcga into path
}
\description{
Download data previously selected using the TCGASeach
}
\examples{
samples <- c("TCGA-06-0125-01A-01D-A45W-05")
   query <- TCGAQuery(tumor = "gbm", platform = "HumanMethylation450",
   level = "3", samples = samples)
   TCGADownload(query,path = "dataDemo2")
}
\seealso{
TCGASearch
}

