% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGAdownload}
\alias{TCGAdownload}
\title{Download the data from TCGA using as reference the output from TCGAquery}
\usage{
TCGAdownload(data = NULL, path = ".", type = NULL, samples = NULL,
  quiet = TRUE, force = FALSE)
}
\arguments{
\item{data}{The TCGAquery output}

\item{path}{Directory to save the downloaded data}

\item{type}{Filter the files that will be downloaded by
type. Example:"rsem.genes.results"}

\item{samples}{List of samples to download data}

\item{quiet}{Supress output messages?. Default: \code{TRUE}}

\item{force}{Download files even if it was already downladed?
Default: \code{FALSE}}
}
\value{
Download TCGA data into the given path
}
\description{
The TCGAdownload function will download the data using as reference
     the the lines of the TCGAquery search result.

     There is an option to download the entire tar.gz folder or download
     specific files using the \emph{type} parameter or the \emph{samples}
     parameter

     The outpufiles will be saved into the path parameters. If this path does
     not exists the package will try to create the directories.

     By default, if a sample was already downloaded the function will not
     download again, unless the force parameter is set to \code{TRUE}
}
\examples{
samples <- c("TCGA-26-1442-01A-01R-1850-01")
query <- TCGAquery(tumor = "gbm",
                   platform = "IlluminaHiSeq_RNASeqV2",
                   level = "3",
                   samples = samples)
TCGAdownload(query,path = "RNA",
             samples = samples,
             type ="rsem.genes.results")
}
\seealso{
\code{\link{TCGAquery}} for searching the data to download

\code{\link{TCGAPrepare}} for preparing the data for the user into
a Summarized experiment object, or a matrix.

Other data.functions: \code{\link{TCGAprepare}};
  \code{\link{TCGAquery}}
}

