% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAintegrate.R
\name{TCGAintegrate}
\alias{TCGAintegrate}
\title{TCGAintegrate}
\usage{
TCGAintegrate(querySamples)
}
\arguments{
\item{querySamples}{output of TCGAQuery filtered}
}
\value{
table with common samples among platforms from TCGAQuery
}
\description{
Common samples among platforms with 2 by 2 comparison
}

