% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAinvestigate.R
\name{TCGAinvestigate}
\alias{TCGAinvestigate}
\title{Find most studied TF in pubmed related to a specific cancer, disease, or tissue}
\usage{
TCGAinvestigate(tumor, dataDEGsFiltLevelTF, topgenes)
}
\arguments{
\item{tumor}{is character such as cancer, disease, or tissue eg. BRCA or breast}

\item{dataDEGsFiltLevelTF}{is a table output from TCGAanalyze_LevelTab with only TFs}

\item{topgenes}{is the number of top genes (eg. 10) in the rownames(dataDEGsFiltLevelTF)
where find in pubmed if those genes or TFs are already related to that cancer or disease}
}
\value{
table with number of pubmed's publications related to tfs and disease selected
}
\description{
Find most studied TF in pubmed related to a specific cancer,
 disease, or tissue
}
\examples{
\dontrun{
TFs <- EAGenes[EAGenes$Family =="transcription regulator",]
TFs_inDEGs <- intersect(TFs$Gene, dataDEGsFiltLevel$mRNA )
dataDEGsFiltLevelTFs <- dataDEGsFiltLevel[TFs_inDEGs,]
dataDEGsFiltLevelTFs <- dataDEGsFiltLevelTFs[order(dataDEGsFiltLevelTFs$Delta,decreasing = TRUE),]
# Find Pubmed of TF studied related to cancer
tabDEGsTFPubmed <- TCGAinvestigate("breast", dataDEGsFiltLevelTFs, topgenes = 1)
}
}

