% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAPrepare}
\alias{TCGAPrepare}
\title{TCGA TCGAPrepare}
\usage{
TCGAPrepare(query, dir = NULL, type = NULL)
}
\value{
Data frame with data read
}
\description{
Prepare data matrices for downstream analysis,
 ready to use also with other R packages
 DNA Methylation
  Row: matrix with genes/loci
  Cols: samples in columns
}
\examples{
TCGAPrepare2(query,dir="path")
}

