% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAPrepare}
\alias{TCGAPrepare}
\title{TCGA Prepare}
\usage{
TCGAPrepare(sdrfFolder = "", downloadFolder = "", PlatformType = "")
}
\arguments{
\item{sdrfFolder}{TCGAQuery location of the output data saving}

\item{downloadFolder}{TCGADownload location of the output data saving}

\item{PlatformType}{Platform as output of TCGAQuery}
}
\value{
Data matrix
}
\description{
Prepare data matrices for downstream analysis, ready to use also with other R packages using data previously downloaded using TCGAdownload
}
\examples{
\dontrun{
   TCGAPrepare(sdrfFolder = 'folder1', downloadFolder= 'folder2')
}
}
\seealso{
TCGAQuery
}

