% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAPrepare}
\alias{TCGAPrepare}
\title{TCGA TCGAPrepare}
\usage{
TCGAPrepare(query, dir = NULL, type = NULL)
}
\arguments{
\item{query}{Data frame as the one returned from TCGAQuery}

\item{dir}{Directory with the files}

\item{type}{File to prepare.}
}
\value{
Data frame with data read
}
\description{
Prepare data matrices for downstream analysis,
 ready to use also with other R packages
 DNA Methylation
  Row: matrix with genes/loci
  Cols: samples in columns
}
\examples{
query <- TCGAQuery(tumor = "gbm") <-
 TCGAPrepare(query,dir="path")
}

