% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAPrepare}
\alias{TCGAPrepare}
\title{TCGA TCGAPrepare}
\usage{
TCGAPrepare(query, dir = NULL, type = NULL)
}
\arguments{
\item{query}{Data frame as the one returned from TCGAQuery}

\item{dir}{Directory with the files}

\item{type}{File to prepare.}
}
\value{
Data frame with data read
}
\description{
Prepare data matrices for downstream analysis,
 ready to use also with other R packages
 DNA Methylation
  Row: matrix with genes/loci
  Cols: samples in columns
}
\examples{
\dontrun{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAQuery(tumor = "GBM",samples = sample, level = 3)
TCGADownload(query,path = "exampleData",samples = sample, quiet = TRUE)
prepared <- TCGAPrepare(query, dir="exampleData")
}
}

