% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAPrepare}
\alias{TCGAPrepare}
\title{TCGA TCGAPrepare}
\usage{
TCGAPrepare(query, dir = NULL, type = NULL, save = FALSE,
  filename = NULL, toPackage = NULL, summarizedExperiment = TRUE)
}
\arguments{
\item{query}{Data frame as the one returned from TCGAQuery}

\item{dir}{Directory with the files}

\item{type}{File to prepare.}

\item{save}{Save a rda object with the prepared object? Default FALSE}

\item{filename}{Name of the saved file}

\item{toPackage}{For whihc package are you preparing the data?}

\item{summarizedExperiment}{Output as SummarizedExperiment? Default: TRUE
Default: NULL. Options: "ELMER"}
}
\value{
A SummarizedExperiment object (If SummarizedExperiment = FALSE,
a data.frame)
}
\description{
Read the data from the experiments and prepare the data for downstream
 analysis.
 The default output is a SummarizedExperiment object.
 It also prepare the data for some specific packages
}
\examples{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAQuery(tumor = "GBM",samples = sample, level = 3)
TCGADownload(query,path = "exampleData",samples = sample, quiet = TRUE)
prepared <- TCGAPrepare(query, dir="exampleData")
}

