% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAPrepare}
\alias{TCGAPrepare}
\title{TCGAPrepare}
\usage{
TCGAPrepare(query, dir = NULL, type = NULL, save = FALSE,
  filename = NULL, toPackage = NULL, summarizedExperiment = TRUE)
}
\arguments{
\item{query}{TCGAQuery output}

\item{dir}{Directory with the files downloaded by TCGADownload}

\item{type}{Filter the files to prepare.}

\item{save}{Save a rda object with the prepared object?
Default: \code{FALSE}}

\item{filename}{Name of the saved file}

\item{toPackage}{For whihc package are you preparing the data?
Default: NULL. Options: "ELMER"}

\item{summarizedExperiment}{Output as SummarizedExperiment?
Default: \code{FALSE}}
}
\value{
A SummarizedExperiment object (If SummarizedExperiment = \code{FALSE},
a data.frame)
}
\description{
This function will read the data from the experiments and prepare it
 for downstream analysis.

 The samples are always refered by their barcode.

 If you want to save the data into an rda file, please use the \emph{save}
 parameter that will save an rda object with the  \emph{filename} parameter.
 If no filename was set, the filename will be the concatenation of platform and
 Sys.time.

 \strong{Return}The default output is a SummarizedExperiment object.
}
\examples{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAQuery(tumor = "GBM",samples = sample, level = 3)
TCGADownload(query,path = "exampleData",samples = sample, quiet = TRUE)
data <- TCGAPrepare(query, dir="exampleData")
}
\seealso{
\code{\link{TCGAQuery}} for searching the data to download

 \code{\link{TCGADownload}} for downloading the data from the
search

Other data.functions: \code{\link{TCGADownload}};
  \code{\link{TCGAQuery}}
}

