% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAprepare}
\alias{TCGAprepare}
\title{Read the data from level 3 the experiments and prepare it
 for downstream analysis into a SummarizedExperiment object.}
\usage{
TCGAprepare(query, dir = NULL, samples = NULL, type = NULL,
  save = FALSE, filename = NULL, reannotate = FALSE,
  summarizedExperiment = TRUE)
}
\arguments{
\item{query}{TCGAquery output}

\item{dir}{Directory with the files downloaded by TCGAdownload}

\item{samples}{List of samples to prepare the data}

\item{type}{Filter the files to prepare.}

\item{save}{Save a rda object with the prepared object?
Default: \code{FALSE}}

\item{filename}{Name of the saved file}

\item{reannotate}{Reannotate genes?  Source http://grch37.ensembl.org/.
DEFAULT: FALSE. (For the moment only working for methylation data)}

\item{summarizedExperiment}{Output as SummarizedExperiment?
Default: \code{FALSE}}
}
\value{
A SummarizedExperiment object (If SummarizedExperiment = \code{FALSE},
a data.frame)
}
\description{
This function will read the data from level 3 the experiments and prepare it
 for downstream analysis into a SummarizedExperiment object.

 The samples are always refered by their barcode.

 If you want to save the data into an rda file, please use the \emph{save}
 parameter that will save an rda object with the  \emph{filename} parameter.
 If no filename was set, the filename will be the concatenation of platform and
 Sys.time.

List of accepted platforms:
\itemize{
\item AgilentG4502A_07_1/AgilentG4502A_07_2/AgilentG4502A_07_3
\item Genome_Wide_SNP_6
\item H-miRNA_8x15K/H-miRNA_8x15Kv2
\item HG-U133_Plus_2
\item HT_HG-U133A
\item HumanMethylation27
\item HumanMethylation450
\item IlluminaDNAMethylation_OMA002_CPI
\item IlluminaDNAMethylation_OMA003_CPI
\item IlluminaGA_RNASeq
\item IlluminaGA_RNASeqV2
\item IlluminaHiSeq_RNASeq
\item IlluminaHiSeq_RNASeqV2
\item IlluminaHiSeq_TotalRNASeqV2
}
 \strong{Return}The default output is a SummarizedExperiment object.
}
\examples{
sample <- "TCGA-06-0939-01A-01D-1228-05"
query <- TCGAquery(tumor = "GBM",samples = sample, level = 3)
TCGAdownload(query,path = "exampleData",samples = sample)
data <- TCGAprepare(query, dir="exampleData")
}
\seealso{
\code{\link{TCGAquery}} for searching the data to download

 \code{\link{TCGAdownload}} for downloading the data from the
search

Other data.functions: \code{\link{TCGAdownload}};
  \code{\link{TCGAquery}}
}

