% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAprepare_added.R
\name{TCGAprepare_Affy}
\alias{TCGAprepare_Affy}
\title{Prepare CEL files into an AffyBatch.}
\usage{
TCGAprepare_Affy(ClinData, PathFolder, TabCel)
}
\arguments{
\item{ClinData}{}

\item{PathFolder}{}

\item{TabCel}{}
}
\value{
Normalizd Expression data from Affy eSets
}
\description{
Prepare CEL files into an AffyBatch.
}
\examples{
\dontrun{
to add example
}
}

