% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAPrepare.R
\name{TCGAprepare_elmer}
\alias{TCGAprepare_elmer}
\title{Prepare the data for ELEMR package}
\usage{
TCGAprepare_elmer(data, platform, met.na.cut = 0.2, save = FALSE)
}
\arguments{
\item{data}{A data frame or summarized experiment from TCGAPrepare}

\item{platform}{platform of the data}

\item{met.na.cut}{Define the percentage of NA that the line should have to
remove the probes for humanmethylation platforms.}

\item{save}{Save object? Default: FALSE.
Names of the files will be: "Exp_elmer.rda" and   "Met_elmer.rda"}
}
\value{
Matrix prepared for fetch.mee function
}
\description{
Prepare the data for ELEMR package
}
\examples{
df <- data.frame(runif(200, 1e5, 1e6),runif(200, 1e5, 1e6))
rownames(df) <- sprintf("?|\%03d", 1:200)
TCGAprepare_elmer(df,platform="IlluminaHiSeq_RNASeqV2")
}

