% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGApreprocessing.R
\name{TCGApreprocessing}
\alias{TCGApreprocessing}
\title{TCGApreprocessing}
\usage{
TCGApreprocessing(object, tabGroupCol)
}
\arguments{
\item{object}{object of class eset or gene expresion}

\item{tabGroupCol}{table with group samples information in tabGroupCol$Color}
}
\value{
Plot with array array intensity correlation and boxplot of correlation samples by samples
}
\description{
Plot with array array intensity correlation and boxplot of correlation samples by samples
}

