% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAQuery.R
\name{tcgaQuery}
\alias{tcgaQuery}
\title{TCGAQuery}
\usage{
tcgaQuery(tumor = NULL, platform = NULL, added.since = NULL,
  added.up.to = NULL, samples = NULL, center = NULL, level = NULL)
}
\arguments{
\item{tumor}{Disease Examples:
\tabular{lllll}{
OV   \tab BRCA \tab CESC \tab ESCA \tab PCPG\cr
LUSC \tab LGG  \tab SKCM \tab KICH \tab CHOL\cr
GBM  \tab UCEC \tab PRAD \tab PAAD \tab THYM\cr
KIRC \tab THCA \tab SARC \tab LAML \tab TGCT\cr
COAD \tab KIRP \tab HNSC \tab ACC  \tab UVM \cr
READ \tab BLCA \tab DLBC \tab UCS  \tab FPPP\cr
LUAD \tab LIHC \tab STAD \tab MESO \tab CNTL
}}

\item{platform}{Example:
\tabular{ll}{
CGH- 1x1M_G4447A                   \tab IlluminaGA_RNASeqV2   \cr
AgilentG4502A_07                  \tab IlluminaGA_mRNA_DGE    \cr
Human1MDuo                        \tab HumanMethylation450    \cr
HG-CGH-415K_G4124A                \tab IlluminaGA_miRNASeq    \cr
HumanHap550                       \tab IlluminaHiSeq_miRNASeq \cr
ABI                               \tab H-miRNA_8x15K  \cr
HG-CGH-244A                       \tab SOLiD_DNASeq                       \cr
IlluminaDNAMethylation_OMA003_CPI \tab IlluminaGA_DNASeq_automated        \cr
IlluminaDNAMethylation_OMA002_CPI \tab HG-U133_Plus_2                 \cr
HuEx- 1_0-st-v2 \tab Mixed_DNASeq                  \cr
H-miRNA_8x15Kv2 \tab IlluminaGA_DNASeq_curated      \cr
MDA_RPPA_Core   \tab IlluminaHiSeq_TotalRNASeqV2    \cr
HT_HG-U133A     \tab IlluminaHiSeq_DNASeq_automated \cr
diagnostic_images                 \tab microsat_i                     \cr
IlluminaHiSeq_RNASeq              \tab SOLiD_DNASeq_curated           \cr
IlluminaHiSeq_DNASeqC             \tab Mixed_DNASeq_curated           \cr
IlluminaGA_RNASeq                 \tab IlluminaGA_DNASeq_Cont_automated   \cr
IlluminaGA_DNASeq                 \tab IlluminaHiSeq_WGBS             \cr
pathology_reports                 \tab IlluminaHiSeq_DNASeq_Cont_automated\cr
Genome_Wide_SNP_6                 \tab bio                            \cr
tissue_images                     \tab Mixed_DNASeq_automated         \cr
HumanMethylation27                \tab Mixed_DNASeq_Cont_curated      \cr
IlluminaHiSeq_RNASeqV2            \tab Mixed_DNASeq_Cont
}}

\item{added.since}{04- 14-2010}

\item{added.up.to}{04- 14-2010}

\item{samples}{List of samples. Ex:c('TCGA-04-06-*','TCGA-04-08-*')}

\item{center}{center name}

\item{level}{'1' '2' '3'}
}
\value{
A dataframe with the results of the query
       (lastest version of the files)
}
\description{
Searches in the tcga database
}
\examples{
query <- tcgaQuery(tumor = "gbm",
                    added.since = "01/01/2013",
                    added.up.to = "06/01/2013")

query <- tcgaQuery(tumor = c("gbm","lgg"),
                    platform = c("HumanMethylation450","HumanMethylation27"))

query <- tcgaQuery(tumor = "gbm", platform = "bio")

query <- tcgaQuery(tumor = "gbm", platform = "HumanMethylation450",
                    level = "3")

query <- tcgaQuery(samples = "TCGA-61-1743-01A-01D-*")

query <- tcgaQuery(samples = "TCGA-61-1743-01A-01D-0649-04")

query <- tcgaQuery(samples = "TCGA-61-1743-01A-01D-0649-04", level = 3)

query <- tcgaQuery(samples = "TCGA-61-1743-01A-01D-0649-04",
                    tumor = "OV", platform = "CGH-1x1M_G4447A")
}

