% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAintegrate.R
\name{TCGAquery_samplesfilter}
\alias{TCGAquery_samplesfilter}
\title{Filtering sample output from TCGAquery}
\usage{
TCGAquery_samplesfilter(query)
}
\arguments{
\item{query}{metaData output from TCGAquery}
}
\value{
list of samples for a tumor
}
\description{
Filtering sample output from TCGAquery
}
\examples{
query <- TCGAquery(tumor = 'brca',level = 3)
querySamples <- TCGAquery_samplesfilter(query)
}

