% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAquery_subtypes.R
\name{TCGAquery_subtypes}
\alias{TCGAquery_subtypes}
\title{Retrieve multiple molecular subtypes according last publications}
\usage{
TCGAquery_subtypes(tumor = NULL, path = ".")
}
\arguments{
\item{tumor}{is character such as cancer from TCGA. Examples:
\tabular{lllll}{
blca \tab gbm  \tab hnsc \tab kich \tab kirc\cr
laml \tab lgg  \tab luad \tab skcm \tab stad\cr
thca \tab ucec
}}

\item{path}{Directory to save the downloaded data}
}
\value{
data.frame with information about molecular cancer subtypes
}
\description{
Retrieve multiple molecular subtypes according last publications
}
\examples{
\dontrun{
GBM_path_subtypes <- TCGAquery_subtypes(tumor = "gbm",path ="dataGBM")
# to prepare run:
# require(xlsx)
# GBM_subtypes <- read.xlsx2(GBM_path_subtypes,1,stringsAsFactors = NULL)
}
}

