% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAsocial.R
\name{TCGAsocial}
\alias{TCGAsocial}
\title{TCGAsocial}
\usage{
TCGAsocial(siteToFind, listPackage = NULL, KeyInfo = NULL)
}
\arguments{
\item{siteToFind}{website related to TCGA or a package}

\item{listPackage}{list of package to find}

\item{KeyInfo}{is a key to find related to TGGA or package}
}
\value{
table with number of downloads about a package
}
\description{
Find number of downloads about a package in CRAN or BioC
}
\examples{
TCGAsocial("bioconductor.org","BiocCheck")
}

