% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAupdate.R
\name{TCGAUpdate}
\alias{TCGAUpdate}
\title{TCGA TCGAUpdate}
\usage{
TCGAUpdate(update = "all")
}
\arguments{
\item{update}{Update options: all (default) disease center platform}
}
\value{
platform.table center.table disease.table tcga.db
}
\description{
Updates local TCGA database
}
\examples{
\dontrun{
  # Update all databases
  TCGAUpdate()
  # Update disease table
  TCGAUpdate(update = "disease")
}
TCGAUpdate("none")
}

