% Generated by roxygen2 (4.1.0): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/TCGAupdate.R, R/TCGAupdate2.R
=======
% Please edit documentation in R/TCGAupdate.R
>>>>>>> master
\name{TCGAUpdate}
\alias{TCGAUpdate}
\title{TCGAUpdate}
\usage{
TCGAUpdate()
}
\description{
replace the dataFolders.rda matrix with the
            new matrix from "The Cancer Genome Atlas" ftp
<<<<<<< HEAD

replace the dataFolders.rda matrix with the
            new matrix from "The Cancer Genome Atlas" ftp
}
\author{
Davide

Davide
}
\seealso{
TCGAQuery

TCGAQuery
=======
}
\author{
Davide
}
\seealso{
TCGAQuery
>>>>>>> master
}

