% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAupdate.R
\name{TCGAUpdate}
\alias{TCGAUpdate}
\title{TCGAUpdate}
\usage{
TCGAUpdate()
}
\description{
replace the dataFolders.rda matrix with the
            new matrix from "The Cancer Genome Atlas" ftp
}
\author{
Davide
}
\seealso{
TCGAQuery
}

