% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAupdate.R
\name{TCGAUpdate}
\alias{TCGAUpdate}
\title{TCGAUpdate}
\usage{
TCGAUpdate()
}
\value{
platform/center/disease/data tables will be updated in the package
}
\description{
Updates the preprocessed TCGA database used by the package.

This function can update the the disease, platform, center, data table.
Updating disease, platform, center will take some seconds. And Updating the
data.table shoul take no more than 10 minutes.

The package will be updated with lastest version of the table every week.
}
\keyword{internal}

