% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGA_biolinks_functions.R
\name{TCGAVersion}
\alias{TCGAVersion}
\title{TCGA Download}
\usage{
TCGAVersion(Tumor, PlatformType, PlatformAndAssociatedData)
}
\arguments{
\item{Tumor}{}

\item{PlatformType}{}

\item{PlatformAndAssociatedData}{}
}
\description{
TCGA Download
}

