% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAVersion.R
\name{TCGAVersion}
\alias{TCGAVersion}
\title{TCGA Version Detailed}
\usage{
TCGAVersion(tumor = "all", centerType = "all", center = "all",
  platform = "all", level = "all", version = "all", barcode = F,
  file = system.file("extdata/dataFolders.rda", package = "TCGAbiolinks"),
  qOutput = "data/version/")
}
\arguments{
\item{tumor}{tumor code between "acc"  "blca" "brca" "cesc" "chol" "cntl" "coad" "dlbc" "esca" "fppp" "gbm"
                                "hnsc" "kich" "kirc" "kirp" "laml" "lcml" "lgg"  "lihc" "lnnh" "luad" "lusc"
                                "meso" "misc" "ov"   "paad" "pcpg" "prad" "read" "sarc" "skcm" "stad" "tgct"
                                "thca" "thym" "ucec" "ucs"  "uvm"}

\item{centerType}{type code between "bcr"  "cgcc" "gsc"}

\item{center}{center code between "biotab"                  "nationwidechildrens.org"
                                  "bcgsc.ca"                "broad.mit.edu"
                                  "jhu-usc.edu"             "mdanderson.org"
                                  "unc.edu"                 "hgsc.bcm.edu"
                                  "hms.harvard.edu"         "genome.wustl.edu"
                                  "ucsc.edu"                "intgen.org"
                                  "hudsonalpha.org"         "lbl.gov"
                                  "mskcc.org"               "supplemental"}

\item{platform}{platform code between  "clin"                                "bio"
                                       "biotab"                              "diagnostic_images"
                                       "pathology_reports"                   "tissue_images"
                                       "illuminahiseq_mirnaseq"              "genome_wide_snp_6"
                                       "humanmethylation450"                 "mda_rppa_core"
                                       "illuminahiseq_rnaseqv2"              "illuminaga_dnaseq_curated"
                                       "illuminaga_dnaseq_automated"         "illuminaga_dnaseq_cont_automated"
                                       "mixed_dnaseq_curated"                "illuminaga_mirnaseq"
                                       "illuminahiseq_dnaseqc"               "illuminahiseq_wgbs"
                                       "illuminahiseq_rnaseq"                "illuminahiseq_totalrnaseqv2"
                                       "illuminaga_dnaseq"                   "humanmethylation27"
                                       "agilentg4502a_07_3"                  "illuminahiseq_dnaseq_automated"
                                       "illuminahiseq_dnaseq_cont_automated" "miRNASeq"
                                       "microsat_i"                          "illuminaga_rnaseq"
                                       "illuminaga_rnaseqv2"                 "RNASeq"
                                       "solid_dnaseq"                        "mixed_dnaseq_automated"
                                       "minbio"                              "abi"
                                       "ht_hg-u133a"                         "hg-cgh-244a"
                                       "hg-cgh-415k_g4124a"                  "humanhap550"
                                       "illuminadnamethylation_oma002_cpi"   "illuminadnamethylation_oma003_cpi"
                                       "huex-1_0-st-v2"                      "agilentg4502a_07_1"
                                       "agilentg4502a_07_2"                  "h-mirna_8x15k"
                                       "h-mirna_8x15kv2"                     "mixed_dnaseq_cont"
                                       "mixed_dnaseq"                        "mixed_dnaseq_cont_curated"
                                       "hg-u133_plus_2"                      " "
                                       "human1mduo"                          "cgh-1x1m_g4447a"
                                       "illuminaga_mrna_dge"                 "solid_dnaseq_curated"}

\item{level}{level 1 2 3}

<<<<<<< Updated upstream
\item{version}{version code -TO DO-}

=======
>>>>>>> Stashed changes
\item{barcode}{Get for each version all barcodes?
Default FALSE
Takes a lot of time to get barcode}

<<<<<<< Updated upstream
=======
\item{version}{version code -TO DO-}

>>>>>>> Stashed changes
\item{file}{- link reference data matrix}

\item{qOutput}{place where the query is saved to be downloaded automatically.
       The folder can be specified in both TCGAQuery and TCGADownload}
}
\description{
TCGA Version Detailed
}
\author{
Tiago
}

