% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_Heatmap}
\alias{TCGAvisualize_Heatmap}
\title{Heatmap with more sensible behavior using heatmap.plus}
\usage{
TCGAvisualize_Heatmap(data, col.metadata, row.metadata, sortCol = NULL,
  sortRow = NULL, col.labels = NULL, row.labels = NULL,
  col.colors = NULL, row.colors = NULL, type = "expression",
  filename = "Heatmap.pdf")
}
\arguments{
\item{data}{The object to with the heatmap data (expression, methylation)}

\item{col.metadata}{Dataframe with the col.label and rownames with patients ids}

\item{row.metadata}{Dataframe with the row.label and rownames as genes (expression) or probes (methylation)}

\item{sortCol}{column used for sorting}

\item{sortRow}{column used for sorting}

\item{col.labels}{Vector of columns to add to the heatpmap col labels}

\item{row.labels}{Vector of columns to add to the heatpmap row labels}

\item{col.colors}{A list of colors, aech one will be used in the labelCols}

\item{row.colors}{A list of colors, aech one will be used in the labelRows}

\item{type}{Select the colors of the heatmap values.
Possible values "expression" (default), "methylation"}

\item{filename}{Filename default "Heatmap.pdf"}
}
\value{
Heatmap plotted in pdf or png file.
}
\description{
Heatmap with more sensible behavior using heatmap.plus
}
\examples{
dat <- matrix(c(0.3,0.2,0.3,
                1,1,0.1,
                1,1,0,
                0.8,1,0.7,
                0.7,0.3,1), nrow = 5, ncol = 3, byrow = TRUE,
               dimnames = list(
               c("probe1", "probe2","probe3","probe4","probe5"),
               c("TCGA-DU-6410",
                 "TCGA-DU-A5TS",
                 "TCGA-HT-7688")
             ))

mdat <- matrix(c("Male","coc1","IDHwt",
                "Male","coc1","IDHMut-cod",
                "Famele","coc1","IDHMut-noncod"),
              nrow = 3, ncol = 3, byrow = TRUE,
              dimnames = list(
                  c("TCGA-DU-6410",
                    "TCGA-DU-A5TS",
                    "TCGA-HT-7688"),
                  c("Sex", "COCCluster","IDHtype")))

TCGAvisualize_Heatmap(dat,mdat,col.labels = c("Sex","COCCluster","IDHtype"),
                     filename = "a.pdf",
                     col.colors = list(c("green","pink"),
                                      c("grey","black","purple"),
                                     c("cyan","tomato","gold")),
                                     type = "methylation")
\dontrun{
# from case study n.2 LGG to test the function
TCGAvisualize_Heatmap(datFilt,
                      clin_subt,
                      col.labels = c("histological_type",
                                     "IDH.1p19q.Subtype",
                                     "CNCluster",
                                     "COCCluster",
                                     "OncosignCluster",
                                     "groupsHC),
                      filename = "a.png",
                      col.colors = list(c("cyan","green3","red","purple"),
                                       c("cyan","tomato","gold"),
                                       c("green","red","purple"),
                                       c("green","red","purple"),
                                       c("green","red","purple","orange","gray"),
                                       c("grey","purple","blue","red"))
                     )
}
}

