% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_Heatmap}
\alias{TCGAvisualize_Heatmap}
\title{Heatmap with more sensible behavior using heatmap.plus}
\usage{
TCGAvisualize_Heatmap(data, col.metadata, row.metadata, col.colors = NULL,
  row.colors = NULL, show_column_names = FALSE, show_row_names = FALSE,
  cluster_rows = FALSE, cluster_columns = FALSE, type = "expression")
}
\arguments{
\item{data}{The object to with the heatmap data (expression, methylation)}

\item{col.metadata}{Metadata for the columns (patients)}

\item{row.metadata}{Metadata for the rows  genes (expression) or probes (methylation)}

\item{col.colors}{A list of names colors}

\item{row.colors}{A list of named colors}

\item{show_column_names}{Show column names names? Dafault: FALSE}

\item{show_row_names}{Show row names? Dafault: FALSE}

\item{cluster_rows}{Cluster rows ? Dafault: FALSE}

\item{cluster_columns}{Cluster columns ? Dafault: FALSE}

\item{type}{Select the colors of the heatmap values. Possible values are
"expression" (default), "methylation"}
}
\value{
Heatmap plotted in pdf or png file.
}
\description{
Heatmap with more sensible behavior using heatmap.plus
}
\examples{
 row.mdat <- matrix(c("FALSE","FALSE",
                     "TRUE","TRUE",
                     "FALSE","FALSE",
                     "TRUE","FALSE",
                     "FALSE","TRUE"
                ),
              nrow = 5, ncol = 2, byrow = TRUE,
              dimnames = list(
                  c("probe1", "probe2","probe3","probe4","probe5"),
                  c("duplicated", "Enhancer region")))
dat <- matrix(c(0.3,0.2,0.3,1,1,0.1,1,1,0, 0.8,1,0.7,0.7,0.3,1),
             nrow = 5, ncol = 3, byrow = TRUE,
               dimnames = list(
               c("probe1", "probe2","probe3","probe4","probe5"),
               c("TCGA-DU-6410",
                 "TCGA-DU-A5TS",
                 "TCGA-HT-7688")))

mdat <- matrix(c("Male","coc1","IDHwt",
                "Male","coc1","IDHMut-cod",
                "Famele","coc1","IDHMut-noncod"),
              nrow = 3, ncol = 3, byrow = TRUE,
              dimnames = list(
                  c("TCGA-DU-6410",
                    "TCGA-DU-A5TS",
                    "TCGA-HT-7688"),
                  c("Sex", "COCCluster","IDHtype")))

TCGAvisualize_Heatmap(dat,
                    col.metadata = mdat,
                    row.metadata = row.mdat,
                    row.colors = list(duplicated = c("FALSE" = "pink",
                                                     "TRUE"="green"),
                                     "Enhancer region" = c("FALSE" = "purple",
                                                            "TRUE"="grey")),
                    col.colors = list(Sex = c("Male" = "blue", "Famele"="red"),
                                      COCCluster=c("coc1"="grey"),
                                      IDHtype=c("IDHwt"="cyan",
                                      "IDHMut-cod"="tomato"
                                      ,"IDHMut-noncod"="gold")),
                    type = "methylation",
                    show_row_names=T)
}

