% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAvisualize_added.R
\name{TCGAvisualize_Heatmap}
\alias{TCGAvisualize_Heatmap}
\title{Heatmap with more sensible behavior using heatmap.plus}
\usage{
TCGAvisualize_Heatmap(DFfilt, DFclin, DFsubt, data_Hc2, cbPalette,
  filename = NULL)
}
\arguments{
\item{DFfilt}{write}

\item{DFclin}{write}

\item{DFsubt}{write}

\item{data_Hc2}{write}

\item{cbPalette}{write}

\item{filename}{write. default = NULL}

\item{Subtype}{write}
}
\value{
Heatmap plotted in pdf or png file.
}
\description{
Heatmap with more sensible behavior using heatmap.plus
}
\examples{
\dontrun{
# from case study n.2 LGG to test the function
DFfilt <- datFilt
DFclin = dataClin
DFsubt = dataSubt
data_Hc2 = data_Hc2
TCGAvisualize_Heatmap(DFfilt,
DFclin,
DFsubt,
data_Hc2)
}
}

