% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGAvisualize.R
\name{TCGAvisualize_profilePlot}
\alias{TCGAvisualize_profilePlot}
\title{Profile plot}
\usage{
TCGAvisualize_profilePlot(data = NULL, groupCol = NULL, subtypeCol = NULL,
  colors = NULL, filename = NULL, na.rm = FALSE, clusterLabel = NULL,
  plot.margin = c(-2.5, -2.5, -0.5, 2), axis.title.size = 1.5,
  axis.textsize = 1.3, legend.size = 1.5, legend.title.size = 1.5,
  geom.label.size = 6)
}
\arguments{
\item{data}{A data frame with the cluters and subytpe of cancers}

\item{groupCol}{Names of tre columns with the cluster information}

\item{subtypeCol}{Name of the column with the subtype information}

\item{colors}{Vector of colors to be used in the bars}

\item{filename}{Name of the file to save the plot, can be pdf, png, svg etc..}

\item{na.rm}{Remove NA groups/subtypes? Default = FALSE}

\item{plot.margin}{Plot margin for cluster distribution. This can control the size
of the bar if the output is not aligned}

\item{axis.title.size}{axis.title.size}

\item{axis.textsize}{axis.textsize}

\item{legend.size}{legend.size}

\item{legend.title.size}{legend.title.size}

\item{geom.label.size}{geom.label.size}
}
\value{
A plot
}
\description{
Displaty the association between cancer subtypes and any kind of clustering.
}
\examples{
cluster <- c(rep("cluster1",30),
             rep("cluster2",30),
             rep("cluster3",30))
subtype <- rep(c(rep("subtype1",10),
           rep("subtype2",10),
           rep("subtype3",10)),3)
df <- data.frame(cluster,subtype)
TCGAvisualize_profilePlot(df, "cluster","subtype")
}

