% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{TCGAvisualize_starburst}
\alias{TCGAvisualize_starburst}
\title{Create starburst plot}
\usage{
TCGAvisualize_starburst(met, exp, group1 = NULL, group2 = NULL,
  exp.p.cut = 0.01, met.p.cut = 0.01, diffmean.cut = 0, logFC.cut = 0,
  filename = "starburst.pdf", ylab = expression(atop("Gene Expression",
  paste(Log[10], " (FDR corrected P values)"))),
  xlab = expression(atop("DNA Methylation", paste(Log[10],
  " (FDR corrected P values)"))), title = "Starburst Plot",
  legend = "Methylation/Expression Relation", color = c(`1` = "black", `2` =
  "coral", `3` = "darkgreen", `4` = "blue", `5` = "darkred", `6` = "red", `7` =
  "brown", `8` = "yellow", `9` = "orange"), label = c(`1` = "Not Significant",
  `2` = "Up regulated & Hypo methylated", `3` =
  "Down regulated & Hypo methylated", `4` = "hypo methylated", `5` =
  "hyper methylated", `6` = "Up regulated", `7` = "Down regulated", `8` =
  "Up regulated & Hyper methylated", `9` = "Down regulated & Hyper methylated"),
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{met}{SummarizedExperiment with methylation data obtained from the
TCGAPrepare. Expected colData columns: diffmean,  p.value.adj  and p.value
Execute volcanoPlot function in order to obtain these values for the object.}

\item{exp}{Object obtained by DEArnaSEQ function}

\item{group1}{The name of the group 1
Obs: Column p.value.adj.group1.group2 should exist}

\item{group2}{The name of the group 2.
Obs: Column p.value.adj.group1.group2 should exist}

\item{exp.p.cut}{expression p value cut-off}

\item{met.p.cut}{methylation p value cut-off}

\item{diffmean.cut}{If set, the probes with diffmean higher than methylation cut-off will be
highlighted in the plot. And the data frame return will be subseted.}

\item{logFC.cut}{If set, the probes with expression fold change higher than methylation cut-off will be
highlighted in the plot. And the data frame return will be subseted.}

\item{filename}{pdf filename}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title}

\item{legend}{legend title}

\item{color}{vector of colors to be used in graph}

\item{label}{vector of labels to be used in graph}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}
}
\value{
Save a starburst plot
}
\description{
Create Starburst plot for comparison of DNA methylation and gene expression.
   The log10 (FDR-corrected P value) is plotted for beta value for DNA
   methylation (x axis) and gene expression (y axis) for each gene.
   The black dashed line shows the FDR-adjusted P value of 0.01.
}
\details{
Input: data with gene expression/methylation expression
   Output: starburst plot
}
\examples{
nrows <- 20000; ncols <- 20
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
ranges <- GenomicRanges::GRanges(rep(c("chr1", "chr2"), c(5000, 15000)),
                   IRanges::IRanges(floor(runif(20000, 1e5, 1e6)), width=100),
                    strand=sample(c("+", "-"), 20000, TRUE),
                    probeID=sprintf("ID\%03d", 1:20000),
                    Gene_Symbol=sprintf("ID\%03d", 1:20000))
colData <- S4Vectors::DataFrame(Treatment=rep(c("ChIP", "Input"), 5),
                    row.names=LETTERS[1:20],
                    group=rep(c("group1","group2"),c(10,10)))
data <- SummarizedExperiment::SummarizedExperiment(
         assays=S4Vectors::SimpleList(counts=counts),
         rowRanges=ranges,
         colData=colData)
met <- data
exp <- data.frame(row.names=sprintf("ID\%03d", 1:20000),
                  logFC=runif(20000, -0.2, 0.2),
                  FDR=runif(20000, 0.01, 1))
SummarizedExperiment::rowRanges(met)$diffmean.g1.g2 <- c(runif(20000, -0.1, 0.1))
SummarizedExperiment::rowRanges(met)$p.value.g1.g2 <- c(runif(20000, 0, 1))
SummarizedExperiment::rowRanges(met)$p.value.adj.g1.g2 <- c(runif(20000, 0, 1))
result <- TCGAvisualize_starburst(met,exp,exp.p.cut = 0.05,met.p.cut = 0.05,"g1","g2")
}

