% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{TCGAVisualize_volcano}
\alias{TCGAVisualize_volcano}
\title{Creates a volcano plot for DNA methylation or expression}
\usage{
TCGAVisualize_volcano(x, y, filename = "volcano.pdf",
  ylab = expression(paste(-Log[10], " (FDR corrected -P values)")),
  xlab = NULL, title = "Volcano plot", legend = NULL, label = NULL,
  xlim = NULL, ylim = NULL, color = c("black", "red", "green"),
  names = NULL, names.fill = TRUE, show.names = "significant",
  x.cut = 0, y.cut = 0.01, height = 5, width = 10, highlight = NULL,
  highlight.color = "orange", names.size = 4, dpi = 300)
}
\arguments{
\item{x}{x-axis data}

\item{y}{y-axis data}

\item{filename}{Filename. Default: volcano.pdf, volcano.svg, volcano.png}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title. If not specified it will be
"Volcano plot (group1 vs group2)}

\item{legend}{Legend title}

\item{label}{vector of labels to be used in the figure.
Example: c("Not Significant","Hypermethylated in group1",
"Hypomethylated in group1"))#'}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}

\item{color}{vector of colors to be used in graph}

\item{names}{Names to be ploted if significant.
Should be the same size of x and y}

\item{names.fill}{Names should be filled in a color box?  Default: TRUE}

\item{show.names}{What names will be showd? Possibilities: "both", "significant", "highlighted"}

\item{x.cut}{x-axis threshold. Default: 0.0 If you give only one number (e.g. 0.2) the cut-offs will be
-0.2 and 0.2. Or you can give diffenrent cutt-ofs as a vector (e.g. c(-0.3,0.4))}

\item{y.cut}{p-values threshold.}

\item{height}{Figure height}

\item{width}{Figure width}

\item{highlight}{List of genes/probes to be highlighted. It should be in the names argument.}

\item{highlight.color}{Color of the points highlighted}

\item{names.size}{Size of the names text}

\item{dpi}{Figure dpi}
}
\value{
Saves the volcano plot in the current folder
}
\description{
Creates a volcano plot from the
expression and methylation analysis.
}
\details{
Creates a volcano plot from the expression and methylation analysis.
   Please see the vignette for more information
   Observation: This function automatically is called by TCGAanalyse_DMR
}
\examples{
x <- runif(200, -1, 1)
y <- runif(200, 0.01, 1)
TCGAVisualize_volcano(x,y)
TCGAVisualize_volcano(x,y,filename = NULL,y.cut = 10000000,x.cut=0.8,
                      names = rep("AAAA",length(x)), legend = "Status",
                      names.fill = FALSE)
TCGAVisualize_volcano(x,y,filename = NULL,y.cut = 10000000,x.cut=0.8,
                      names = as.character(1:length(x)), legend = "Status",
                      names.fill = TRUE, highlight = c("1","2"),show="both")
TCGAVisualize_volcano(x,y,filename = NULL,y.cut = 10000000,x.cut=c(-0.3,0.8),
                      names = as.character(1:length(x)), legend = "Status",
                      names.fill = TRUE, highlight = c("1","2"),show="both")
while (!(is.null(dev.list()["RStudioGD"]))){dev.off()}
}
