% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{TCGAVisualize_volcano}
\alias{TCGAVisualize_volcano}
\title{Plot volcano plot for DNA methylation or expression}
\usage{
TCGAVisualize_volcano(x, y, filename = "volcano.pdf",
  ylab = expression(paste(-Log[10], " (FDR corrected -P values)")),
  xlab = NULL, title = NULL, legend = NULL, label = NULL, xlim = NULL,
  ylim = NULL, color = c("black", "red", "darkgreen"), names = NULL,
  x.cut = 0, y.cut = 0.01, height = 5, width = 10)
}
\arguments{
\item{x}{x-axis data}

\item{y}{y-axis data}

\item{filename}{Filename. Default: volcano.pdf, volcano.svg, volcano.png}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title. If not specified it will be
"Volcano plot (group1 vs group2)}

\item{legend}{Legend title}

\item{label}{vector of labels to be used in the figure.
Example: c("Not Significant","Hypermethylated in group1",
"Hypomethylated in group1"))#'}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}

\item{color}{vector of colors to be used in graph}

\item{names}{Names to be ploted if significant.
Should be the same size of x and y}

\item{height}{Figure height}

\item{width}{Figure width}
}
\value{
Saves the volcano plot in the current folder
}
\description{
Plot volcano plot for DNA methylation or expression
}
\details{
Creates a volcano plot from the expression and methylation analysis.
   Please see the vignette for more information
   Observation: This function automatically is called by TCGAanalyse_DMR
}
\examples{
x <- runif(200, 1e5, 1e6)
y <- runif(200, 1e5, 1e6)
TCGAVisualize_volcano(x,y)
}

