% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(data, filename = "volcano.pdf",
  ylab = "- 1*log10 of the Significance", xlab = "DNA Methylation",
  title = "Volcano Plot TCGA GBM Tumors", legend = "Legend", color = c(`1`
  = "black", `2` = "green", `3` = "red"), label = c(`1` = "Not Significant",
  `2` = "Hypermethylated", `3` = "Hypomethylated"), xlim = NULL,
  ylim = NULL, p.cut = 0.05, diffmean.cut = 0)
}
\arguments{
\item{data}{data frame columns: diffmean, p.value.adj}

\item{filename}{pdf filename}

\item{ylab}{y axis text}

\item{xlab}{x axis text}

\item{title}{main title}

\item{legend}{legend title}

\item{color}{vector of colors to be used in graph}

\item{label}{vector of labels to be used in the figure}

\item{xlim}{x limits to cut image}

\item{ylim}{y limits to cut image}

\item{p.cut}{p values threshold}

\item{diffmean.cut}{diffmean threshold}
}
\value{
A dataframe with the Composite.Element.REF and
        the group it was classified
         group 1 = Not Significant
         group 2 = Hypermethylated
         group 3 = Hypomethylated
}
\description{
Volcano plot - hypo/hyper methylation graphic.
  Input: a dataframe with columns:
   1 - p.value.adj - p value adjusted
   2 - diffmean - difference between old stated and new state of methylation
}
\examples{
patient <- paste("patient",1:100)
Composite.Element.REF <- paste0("cg00",1:100)
beta.values <- runif(10000,0,1)
beta.values <- matrix(beta.values,ncol = 100,nrow = 100)
values <- as.data.frame(beta.values)
colnames(values) <- patient
rownames(values) <- Composite.Element.REF
data <- calculate.pvalues(as.data.frame(t(values)),1:50,51:100)
diffmean <- diffmean(values[,1:50],values[,51:100])
data <-cbind(Composite.Element.REF,data,diffmean)
hypo.hyper <- volcanoPlot(data, p.cut = 0.85)
}

